@@ -1,115 +1,131 @@
 {
     "name": "fuse-react-app",
-    "version": "7.1.0",
+    "version": "8.3.7",
     "private": true,
     "dependencies": {
-        "@emotion/cache": "11.7.1",
-        "@emotion/react": "11.7.1",
-        "@emotion/styled": "11.6.0",
-        "@hookform/resolvers": "2.8.5",
-        "@mui/base": "5.0.0-alpha.60",
-        "@mui/icons-material": "5.2.4",
-        "@mui/lab": "5.0.0-alpha.60",
-        "@mui/material": "5.2.4",
-        "@mui/styles": "5.2.3",
-        "@mui/system": "5.2.4",
-        "@mui/utils": "5.2.3",
-        "@mui/x-data-grid": "5.2.0",
-        "@reduxjs/toolkit": "1.7.0",
+        "@apollo/client": "^3.7.15",
+        "@emotion/cache": "11.11.0",
+        "@emotion/react": "11.11.0",
+        "@emotion/styled": "11.11.0",
+        "@hookform/resolvers": "3.1.0",
+        "@mui/base": "5.0.0-beta.2",
+        "@mui/icons-material": "5.11.16",
+        "@mui/lab": "5.0.0-alpha.131",
+        "@mui/material": "5.13.2",
+        "@mui/material-next": "6.0.0-alpha.86",
+        "@mui/styles": "5.13.2",
+        "@mui/system": "5.13.2",
+        "@mui/utils": "5.13.1",
+        "@mui/x-data-grid": "6.5.0",
+        "@mui/x-date-pickers": "6.5.0",
+        "@reduxjs/toolkit": "1.9.5",
         "@tinymce/tinymce-react": "^4.3.0",
-        "auth0-lock": "11.31.1",
-        "autosuggest-highlight": "3.2.0",
+        "apollo-link-error": "^1.1.13",
+        "apollo-upload-client": "^17.0.0",
+        "apollo-utilities": "^1.3.4",
+        "autosuggest-highlight": "3.3.4",
         "axios": "1.4.0",
-        "clsx": "1.1.1",
-        "date-fns": "2.27.0",
+        "axios-mock-adapter": "1.21.4",
+        "clsx": "1.2.1",
+        "core-js": "3.30.2",
+        "crypto-js": "4.1.1",
+        "date-fns": "2.30.0",
         "draft-js": "0.11.7",
         "draftjs-to-html": "0.9.1",
-        "firebase": "9.6.1",
-        "framer-motion": "5.5.1",
-        "history": "5.1.0",
-        "i18next": "21.6.0",
+        "framer-motion": "10.12.16",
+        "graphql": "^16.6.0",
+        "history": "5.3.0",
+        "i18next": "22.5.0",
         "jwt-decode": "3.1.2",
+        "keycode": "2.2.1",
         "lodash": "4.17.21",
+        "marked": "4.3.0",
+        "material-ui-popup-state": "5.0.8",
         "mobile-detect": "1.4.5",
-        "moment": "2.29.1",
+        "mobx": "^6.9.0",
+        "moment": "2.29.4",
         "notistack": "1.0.6-next.3",
-        "perfect-scrollbar": "1.5.3",
-        "prismjs": "1.25.0",
-        "prop-types": "15.7.2",
-        "qs": "6.10.2",
+        "perfect-scrollbar": "1.5.5",
+        "prismjs": "1.29.0",
+        "prop-types": "15.8.1",
+        "qs": "6.11.2",
         "query-string": "^8.1.0",
-        "react": "17.0.2",
+        "react": "18.2.0",
         "react-autosuggest": "10.1.0",
-        "react-dom": "17.0.2",
-        "react-draft-wysiwyg": "1.14.7",
-        "react-hook-form": "7.22.0",
-        "react-i18next": "11.15.1",
-        "react-popper": "1.3.11",
-        "react-redux": "7.2.6",
-        "react-router-dom": "6.1.1",
-        "react-scripts": "5.0.0",
-        "react-swipeable": "6.2.0",
+        "react-dom": "18.2.0",
+        "react-draft-wysiwyg": "1.15.0",
+        "react-hook-form": "7.44.2",
+        "react-i18next": "12.3.1",
+        "react-popper": "2.3.0",
+        "react-redux": "8.0.5",
+        "react-router-dom": "6.11.2",
+        "react-swipeable": "7.0.0",
         "redux-logger": "4.0.0",
-        "styled-components": "5.3.3",
-        "stylis": "4.0.10",
+        "styled-components": "5.3.11",
+        "stylis": "4.2.0",
         "stylis-plugin-rtl": "2.1.1",
-        "web-vitals": "2.1.2",
-        "yup": "0.32.11"
+        "subscriptions-transport-ws": "^0.11.0",
+        "web-vitals": "3.3.1",
+        "yup": "1.2.0"
     },
     "peerDependencies": {
-        "autoprefixer": "9.8.8",
-        "postcss": "8.4.5",
-        "react": "17.0.2",
-        "react-dom": "17.0.2"
+        "autoprefixer": "10.4.7",
+        "postcss": "8.4.20",
+        "react": "18.2.0",
+        "react-dom": "18.2.0"
     },
     "resolutions": {
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
         "babel-loader": "8.1.0"
     },
     "devDependencies": {
-        "@babel/core": "7.15.8",
-        "@babel/node": "7.15.8",
-        "@babel/preset-env": "7.15.8",
-        "@craco/craco": "^6.4.3",
-        "@typescript-eslint/eslint-plugin": "^4.33.0",
-        "@typescript-eslint/parser": "^4.33.0",
-        "autoprefixer": "9.8.8",
-        "babel-eslint": "10.1.0",
+        "@babel/core": "7.21.4",
+        "@babel/eslint-parser": "7.21.3",
+        "@babel/node": "7.20.7",
+        "@babel/preset-env": "7.21.4",
+        "@babel/preset-react": "7.18.6",
+        "@tailwindcss/aspect-ratio": "^0.4.2",
+        "@tailwindcss/typography": "^0.5.9",
+        "@typescript-eslint/eslint-plugin": "^5.57.0",
+        "@typescript-eslint/parser": "^5.57.0",
+        "autoprefixer": "10.4.14",
         "cross-env": "7.0.3",
-        "eslint": "7.32.0",
-        "eslint-config-airbnb": "^18.2.1",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-config-react-app": "^6.0.0",
-        "eslint-plugin-flowtype": "^5.10.0",
-        "eslint-plugin-import": "^2.25.2",
-        "eslint-plugin-jsx-a11y": "^6.4.1",
-        "eslint-plugin-prettier": "^4.0.0",
-        "eslint-plugin-react": "^7.26.1",
-        "eslint-plugin-react-hooks": "^4.2.0",
-        "eslint-plugin-unused-imports": "^1.1.5",
-        "js-beautify": "1.14.0",
-        "postcss": "8.4.5",
-        "prettier": "^2.4.1",
-        "promise": "8.1.0",
-        "source-map-explorer": "2.5.2",
-        "tailwindcss": "3.0.5",
-        "tailwindcss-dir": "4.0.0",
-        "typescript": "4.4.4",
-        "util": "^0.12.4"
+        "eslint": "8.37.0",
+        "eslint-config-airbnb": "^19.0.4",
+        "eslint-config-prettier": "^8.8.0",
+        "eslint-config-react-app": "^7.0.1",
+        "eslint-plugin-flowtype": "^8.0.3",
+        "eslint-plugin-import": "^2.27.5",
+        "eslint-plugin-jsx-a11y": "^6.7.1",
+        "eslint-plugin-prettier": "^4.2.1",
+        "eslint-plugin-react": "^7.32.2",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "eslint-plugin-unused-imports": "^2.0.0",
+        "immutable": "4.3.0",
+        "postcss": "8.4.21",
+        "prettier": "^2.8.7",
+        "promise": "8.3.0",
+        "react-app-alias": "2.2.2",
+        "react-app-rewired": "2.2.1",
+        "react-scripts": "5.0.1",
+        "source-map-explorer": "2.5.3",
+        "tailwindcss": "3.3.2",
+        "typescript": "5.0.3",
+        "util": "^0.12.5"
     },
     "scripts": {
-        "start": "craco start",
-        "build": "cross-env GENERATE_SOURCEMAP=false craco build",
-        "test": "craco test --env=node",
-        "eject": "craco eject",
+        "start": "react-app-rewired  start",
+        "build": "cross-env GENERATE_SOURCEMAP=false react-app-rewired build",
+        "test": "react-app-rewired test --env=node",
+        "eject": "react-app-rewired eject",
         "build-docs": "babel-node --presets @babel/preset-env src/app/main/documentation/material-ui-components/build.js",
-        "analyze": "craco build && source-map-explorer 'build/static/js/*.js' --html analyze-result.html",
+        "analyze": "react-app-rewired build && source-map-explorer 'build/static/js/*.js' --html analyze-result.html",
         "lint": "eslint"
     },
     "engines": {
-        "node": ">=14.0.0",
-        "npm": ">=7.0.0"
+        "node": ">=18.0.0",
+        "npm": ">=8.0.0"
     },
     "browserslist": {
         "production": [
