@@ -5,12 +5,14 @@ import './i18n';
 import './styles/app-base.css';
 import './styles/app-components.css';
 import './styles/app-utilities.css';
-import ReactDOM from 'react-dom';
-import App from 'app/App';
+import { createRoot } from 'react-dom/client';
+import App from './app/App';
 import * as serviceWorker from './serviceWorker';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App />);
 
 reportWebVitals();
 
