@@ -1,11 +1,15 @@
 import Dialog from '@mui/material/Dialog';
 import { useDispatch, useSelector } from 'react-redux';
-import { closeDialog } from 'app/store/fuse/dialogSlice';
+import {
+  closeDialog,
+  selectFuseDialogOptions,
+  selectFuseDialogState,
+} from 'app/store/fuse/dialogSlice';
 
 function FuseDialog(props) {
   const dispatch = useDispatch();
-  const state = useSelector(({ fuse }) => fuse.dialog.state);
-  const options = useSelector(({ fuse }) => fuse.dialog.options);
+  const state = useSelector(selectFuseDialogState);
+  const options = useSelector(selectFuseDialogOptions);
 
   return (
     <Dialog
