@@ -1,72 +1,75 @@
 import * as Prism from 'prismjs';
 import PropTypes from 'prop-types';
-import { memo, useEffect, useRef } from 'react';
+import { useEffect, useMemo, useRef } from 'react';
 import './prism-languages';
 import { styled } from '@mui/material/styles';
+import clsx from 'clsx';
 
 function FuseHighlight(props) {
+  const { async, children, className, component: Wrapper } = props;
+
   const domNode = useRef(null);
-  const source = useRef(trimCode());
 
   useEffect(() => {
-    function highlight() {
-      Prism.highlightElement(domNode.current, props.async);
+    if (domNode.current) {
+      Prism.highlightElement(domNode.current, async);
     }
+  }, [children, async]);
 
-    highlight();
-  }, [props.async]);
+  return useMemo(() => {
+    const trimCode = () => {
+      let sourceString = children;
 
-  function trimCode() {
-    let sourceString = props.children;
-
-    if (typeof sourceString === 'object' && sourceString.default) {
-      sourceString = sourceString.default;
-    }
-
-    // Split the source into lines
-    const sourceLines = sourceString.split('\n');
-
-    // Remove the first and the last line of the source
-    // code if they are blank lines. This way, the html
-    // can be formatted properly while using fuse-highlight
-    // component
-    if (!sourceLines[0].trim()) {
-      sourceLines.shift();
-    }
-
-    if (!sourceLines[sourceLines.length - 1].trim()) {
-      sourceLines.pop();
-    }
-
-    // Find the first non-whitespace char index in
-    // the first line of the source code
-    const indexOfFirstChar = sourceLines[0].search(/\S|$/);
-
-    // Generate the trimmed source
-    let sourceRaw = '';
-
-    // Iterate through all the lines
-    sourceLines.forEach((line, index) => {
-      // Trim the beginning white space depending on the index
-      // and concat the source code
-      sourceRaw += line.substr(indexOfFirstChar, line.length);
-
-      // If it's not the last line...
-      if (index !== sourceLines.length - 1) {
-        // Add a line break at the end
-        sourceRaw = `${sourceRaw}\n`;
+      if (typeof sourceString === 'object' && sourceString.default) {
+        sourceString = sourceString.default;
       }
-    });
-    return sourceRaw;
-  }
 
-  const { className, component: Wrapper } = props;
+      // Split the source into lines
+      const sourceLines = sourceString.split('\n');
 
-  return (
-    <Wrapper ref={domNode} className={className}>
-      {source.current}
-    </Wrapper>
-  );
+      // Remove the first and the last line of the source
+      // code if they are blank lines. This way, the html
+      // can be formatted properly while using fuse-highlight
+      // component
+      if (!sourceLines[0].trim()) {
+        sourceLines.shift();
+      }
+
+      if (!sourceLines[sourceLines.length - 1].trim()) {
+        sourceLines.pop();
+      }
+
+      // Find the first non-whitespace char index in
+      // the first line of the source code
+      const indexOfFirstChar = sourceLines[0].search(/\S|$/);
+
+      // Generate the trimmed source
+      let sourceRaw = '';
+
+      // Iterate through all the lines
+      sourceLines.forEach((line, index) => {
+        // Trim the beginning white space depending on the index
+        // and concat the source code
+        sourceRaw += line.substr(indexOfFirstChar, line.length);
+
+        // If it's not the last line...
+        if (index !== sourceLines.length - 1) {
+          // Add a line break at the end
+          sourceRaw = `${sourceRaw}\n`;
+        }
+      });
+      return sourceRaw || '';
+    };
+
+    return (
+      <>
+        <Wrapper ref={domNode} className={clsx('border', className)}>
+          {/* {trimCode()} */}
+          {trimCode()}
+        </Wrapper>
+      </>
+    );
+  }, [children, className]);
 }
 
 FuseHighlight.propTypes = {
@@ -76,4 +79,4 @@ FuseHighlight.defaultProps = {
   component: 'code',
 };
 
-export default memo(styled(FuseHighlight)``);
+export default styled(FuseHighlight)``;
