@@ -1,9 +1,13 @@
 import { useDeepCompareEffect } from '@fuse/hooks';
-import FuseLayouts from '@fuse/layouts/FuseLayouts';
 import _ from '@lodash';
 import AppContext from 'app/AppContext';
-import { generateSettings, setSettings } from 'app/store/fuse/settingsSlice';
-import { memo, useContext, useMemo, useCallback, useRef } from 'react';
+import {
+  generateSettings,
+  selectFuseCurrentSettings,
+  selectFuseDefaultSettings,
+  setSettings,
+} from 'app/store/fuse/settingsSlice';
+import { memo, useCallback, useContext, useMemo, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { matchRoutes, useLocation } from 'react-router-dom';
 import GlobalStyles from '@mui/material/GlobalStyles';
@@ -20,15 +24,15 @@ const inputGlobalStyles = (
         backgroundColor: theme.palette.background.default,
         color: theme.palette.text.primary,
       },
-      'code:not([class*="language-"])': {
+      /*  'code:not([class*="language-"])': {
         color: theme.palette.secondary.dark,
         backgroundColor:
           theme.palette.mode === 'light' ? 'rgba(255, 255, 255, .9)' : 'rgba(0, 0, 0, .9)',
         padding: '2px 3px',
         borderRadius: 2,
         lineHeight: 1.7,
-      },
-      'table.simple tbody tr td': {
+      }, */
+      'table.simple tbody tr th': {
         borderColor: theme.palette.divider,
       },
       'table.simple thead tr th': {
@@ -49,10 +53,13 @@ const inputGlobalStyles = (
           textDecoration: 'none',
         },
       },
-      '[class^="border-"]': {
+      '[class^="border"]': {
         borderColor: theme.palette.divider,
       },
-      '[class*="border-"]': {
+      '[class*="border"]': {
+        borderColor: theme.palette.divider,
+      },
+      '[class*="divide-"] > :not([hidden]) ~ :not([hidden])': {
         borderColor: theme.palette.divider,
       },
       hr: {
@@ -74,9 +81,10 @@ const inputGlobalStyles = (
 );
 
 function FuseLayout(props) {
+  const { layouts } = props;
   const dispatch = useDispatch();
-  const settings = useSelector(({ fuse }) => fuse.settings.current);
-  const defaultSettings = useSelector(({ fuse }) => fuse.settings.defaults);
+  const settings = useSelector(selectFuseCurrentSettings);
+  const defaultSettings = useSelector(selectFuseDefaultSettings);
 
   const appContext = useContext(AppContext);
   const { routes } = appContext;
@@ -127,7 +135,7 @@ function FuseLayout(props) {
 
   // console.warn('::FuseLayout:: rendered');
 
-  const Layout = useMemo(() => FuseLayouts[settings.layout.style], [settings.layout.style]);
+  const Layout = useMemo(() => layouts[settings.layout.style], [layouts, settings.layout.style]);
 
   return _.isEqual(newSettings.current, settings) ? (
     <>
