@@ -1,6 +1,5 @@
 import NavLinkAdapter from '@fuse/core/NavLinkAdapter';
 import { styled } from '@mui/material/styles';
-import Icon from '@mui/material/Icon';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import clsx from 'clsx';
@@ -8,6 +7,7 @@ import PropTypes from 'prop-types';
 import { memo, useMemo } from 'react';
 import withRouter from '@fuse/core/withRouter';
 import FuseNavBadge from '../../FuseNavBadge';
+import FuseSvgIcon from '../../../FuseSvgIcon';
 
 const StyledListItem = styled(ListItem)(({ theme }) => ({
   color: theme.palette.text.primary,
@@ -38,25 +38,27 @@ function FuseNavHorizontalItem(props) {
       <StyledListItem
         button
         component={NavLinkAdapter}
-        to={item.url}
-        activeClassName="active"
-        className={clsx('fuse-list-item')}
+        to={item.url || ''}
+        activeClassName={item.url ? 'active' : ''}
+        className={clsx('fuse-list-item', item.active && 'active')}
         end={item.end}
         role="button"
+        sx={item.sx}
+        disabled={item.disabled}
       >
         {item.icon && (
-          <Icon
-            className={clsx('fuse-list-item-icon text-16 shrink-0', item.iconClass)}
+          <FuseSvgIcon
+            className={clsx('fuse-list-item-icon shrink-0', item.iconClass)}
             color="action"
           >
             {item.icon}
-          </Icon>
+          </FuseSvgIcon>
         )}
 
         <ListItemText
           className="fuse-list-item-text"
           primary={item.title}
-          classes={{ primary: 'text-13 fuse-list-item-text-primary' }}
+          classes={{ primary: 'text-13 fuse-list-item-text-primary truncate' }}
         />
 
         {item.badge && <FuseNavBadge className="ltr:ml-8 rtl:mr-8" badge={item.badge} />}
