@@ -15,6 +15,12 @@ const StyledList = styled(List)(({ theme }) => ({
         theme.palette.mode === 'dark' ? 'rgba(255, 255, 255, 0.06)' : 'rgba(0,0,0,.05)',
     },
   },
+  '& .fuse-list-item-text': {
+    margin: 0,
+  },
+  '& .fuse-list-item-text-primary': {
+    lineHeight: '20px',
+  },
   '&.active-square-list': {
     '& .fuse-list-item, & .active.fuse-list-item': {
       width: '100%',
@@ -41,7 +47,7 @@ function FuseNavVerticalLayout1(props) {
   return (
     <StyledList
       className={clsx(
-        'navigation whitespace-nowrap px-12',
+        'navigation whitespace-nowrap px-12 py-0',
         `active-${active}-list`,
         dense && 'dense',
         className
