@@ -1,33 +1,36 @@
 import NavLinkAdapter from '@fuse/core/NavLinkAdapter';
-import { styled, alpha } from '@mui/material/styles';
+import { alpha, styled } from '@mui/material/styles';
 import Collapse from '@mui/material/Collapse';
-import Icon from '@mui/material/Icon';
 import IconButton from '@mui/material/IconButton';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
-import { useEffect, useState, useMemo } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useLocation } from 'react-router-dom';
+import List from '@mui/material/List';
 import FuseNavBadge from '../../FuseNavBadge';
 import FuseNavItem from '../../FuseNavItem';
+import FuseSvgIcon from '../../../FuseSvgIcon';
 
-const Root = styled('ul')(({ theme, ...props }) => ({
+const Root = styled(List)(({ theme, ...props }) => ({
   padding: 0,
   '&.open': {},
   '& > .fuse-list-item': {
-    height: 40,
+    minHeight: 44,
     width: '100%',
     borderRadius: '6px',
     margin: '0 0 4px 0',
-    paddingRight: 12,
+    paddingRight: 16,
     paddingLeft: props.itempadding > 80 ? 80 : props.itempadding,
+    paddingTop: 10,
+    paddingBottom: 10,
     color: alpha(theme.palette.text.primary, 0.7),
     '&:hover': {
       color: theme.palette.text.primary,
     },
     '& > .fuse-list-item-icon': {
-      marginRight: 12,
+      marginRight: 16,
       color: 'inherit',
     },
   },
@@ -60,7 +63,7 @@ function isUrlInChildren(parent, url) {
 function FuseNavVerticalCollapse(props) {
   const [open, setOpen] = useState(() => needsToBeOpened(props.location, props.item));
   const { item, nestedLevel, onItemClick } = props;
-  const itempadding = nestedLevel > 0 ? 28 + nestedLevel * 16 : 12;
+  const itempadding = nestedLevel > 0 ? 38 + nestedLevel * 16 : 16;
 
   const location = useLocation();
 
@@ -75,7 +78,7 @@ function FuseNavVerticalCollapse(props) {
 
   return useMemo(
     () => (
-      <Root className={clsx(open && 'open')} itempadding={itempadding}>
+      <Root className={clsx(open && 'open')} itempadding={itempadding} sx={item.sx}>
         <ListItem
           component={item.url ? NavLinkAdapter : 'li'}
           button
@@ -84,33 +87,39 @@ function FuseNavVerticalCollapse(props) {
           to={item.url}
           end={item.end}
           role="button"
+          disabled={item.disabled}
         >
           {item.icon && (
-            <Icon
-              className={clsx('fuse-list-item-icon text-20 shrink-0', item.iconClass)}
+            <FuseSvgIcon
+              className={clsx('fuse-list-item-icon shrink-0', item.iconClass)}
               color="action"
             >
               {item.icon}
-            </Icon>
+            </FuseSvgIcon>
           )}
 
           <ListItemText
             className="fuse-list-item-text"
             primary={item.title}
-            classes={{ primary: 'text-13 font-medium' }}
+            secondary={item.subtitle}
+            classes={{
+              primary: 'text-13 font-medium fuse-list-item-text-primary truncate',
+              secondary:
+                'text-11 font-medium fuse-list-item-text-secondary leading-normal truncate',
+            }}
           />
 
           {item.badge && <FuseNavBadge className="mx-4" badge={item.badge} />}
 
           <IconButton
             disableRipple
-            className="w-40 h-40 -mx-12 p-0 focus:bg-transparent hover:bg-transparent"
+            className="w-20 h-20 -mx-12 p-0 focus:bg-transparent hover:bg-transparent"
             onClick={(ev) => ev.preventDefault()}
             size="large"
           >
-            <Icon className="text-16 arrow-icon" color="inherit">
-              {open ? 'expand_less' : 'expand_more'}
-            </Icon>
+            <FuseSvgIcon size={16} className="arrow-icon" color="inherit">
+              {open ? 'heroicons-solid:chevron-down' : 'heroicons-solid:chevron-right'}
+            </FuseSvgIcon>
           </IconButton>
         </ListItem>
 
