@@ -1,33 +1,30 @@
 import NavLinkAdapter from '@fuse/core/NavLinkAdapter';
-import { styled, useTheme, alpha } from '@mui/material/styles';
+import { alpha, styled } from '@mui/material/styles';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import { useMemo } from 'react';
-import { useDispatch } from 'react-redux';
-import useMediaQuery from '@mui/material/useMediaQuery';
-import ListSubheader from '@mui/material/ListSubheader';
+import ListItem from '@mui/material/ListItem';
+import { ListItemText } from '@mui/material';
 import FuseNavItem from '../../FuseNavItem';
 
-const Root = styled(ListSubheader)(({ theme, itempadding, ...props }) => ({
-  height: 40,
+const Root = styled(ListItem)(({ theme, itempadding, ...props }) => ({
+  minminHeight: 44,
   width: '100%',
   borderRadius: '6px',
-  margin: '24px 0 4px 0',
-  paddingRight: 12,
+  margin: '28px 0 0 0',
+  paddingRight: 16,
   paddingLeft: props.itempadding > 80 ? 80 : props.itempadding,
+  paddingTop: 10,
+  paddingBottom: 10,
   color: alpha(theme.palette.text.primary, 0.7),
   fontWeight: 600,
   letterSpacing: '0.025em',
 }));
 
 function FuseNavVerticalGroup(props) {
-  const dispatch = useDispatch();
-
-  const theme = useTheme();
-  const mdDown = useMediaQuery(theme.breakpoints.down('lg'));
   const { item, nestedLevel, onItemClick } = props;
 
-  const itempadding = nestedLevel > 0 ? 28 + nestedLevel * 16 : 12;
+  const itempadding = nestedLevel > 0 ? 38 + nestedLevel * 16 : 16;
 
   return useMemo(
     () => (
@@ -35,14 +32,41 @@ function FuseNavVerticalGroup(props) {
         <Root
           component={item.url ? NavLinkAdapter : 'li'}
           itempadding={itempadding}
-          disableSticky
-          className={clsx('fuse-list-subheader flex items-center', !item.url && 'cursor-default')}
+          className={clsx(
+            'fuse-list-subheader flex items-center  py-10',
+            !item.url && 'cursor-default'
+          )}
           onClick={() => onItemClick && onItemClick(item)}
           to={item.url}
           end={item.end}
           role="button"
+          sx={item.sx}
+          disabled={item.disabled}
         >
-          <span className="fuse-list-subheader-text uppercase text-12">{item.title}</span>
+          <ListItemText
+            className="fuse-list-subheader-text"
+            sx={{
+              margin: 0,
+              '& > .MuiListItemText-primary': {
+                fontSize: 12,
+                color: 'secondary.light',
+                fontWeight: 600,
+                textTransform: 'uppercase',
+                letterSpacing: '.05em',
+                lineHeight: '20px',
+              },
+
+              '& > .MuiListItemText-secondary': {
+                fontSize: 11,
+                color: 'text.disabled',
+                letterSpacing: '.06px',
+                fontWeight: 500,
+                lineHeight: '1.5',
+              },
+            }}
+            primary={item.title}
+            secondary={item.subtitle}
+          />
         </Root>
         {item.children && (
           <>
