@@ -1,15 +1,8 @@
-import { useTheme, ThemeProvider } from '@mui/material/styles';
-import { selectContrastMainTheme } from 'app/store/fuse/settingsSlice';
-import { useSelector } from 'react-redux';
+import clsx from 'clsx';
 
 function FusePageCardedHeader(props) {
-  const theme = useTheme();
-  const contrastTheme = useSelector(selectContrastMainTheme(theme.palette.primary.main));
-
   return (
-    <div className="FusePageCarded-header">
-      {props.header && <ThemeProvider theme={contrastTheme}>{props.header}</ThemeProvider>}
-    </div>
+    <div className={clsx('FusePageCarded-header', 'container')}>{props.header && props.header}</div>
   );
 }
 
