@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 import FuseScrollbars from '@fuse/core/FuseScrollbars';
-import { useTheme, ThemeProvider } from '@mui/material/styles';
+import { ThemeProvider, useTheme } from '@mui/material/styles';
 import { selectContrastMainTheme } from 'app/store/fuse/settingsSlice';
 import clsx from 'clsx';
 
@@ -12,15 +12,7 @@ function FusePageSimpleSidebarContent(props) {
     <FuseScrollbars enable={props.innerScroll}>
       {props.header && (
         <ThemeProvider theme={contrastTheme}>
-          <div
-            className={clsx(
-              'FusePageSimple-sidebarHeader',
-              props.variant,
-              props.sidebarInner && 'FusePageSimple-sidebarHeaderInnerSidebar'
-            )}
-          >
-            {props.header}
-          </div>
+          <div className={clsx('FusePageSimple-sidebarHeader', props.variant)}>{props.header}</div>
         </ThemeProvider>
       )}
 
