@@ -1,6 +1,5 @@
 import ClickAwayListener from '@mui/material/ClickAwayListener';
 import { styled } from '@mui/material/styles';
-import Icon from '@mui/material/Icon';
 import IconButton from '@mui/material/IconButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
@@ -16,9 +15,8 @@ import clsx from 'clsx';
 import _ from '@lodash';
 import { memo, useEffect, useReducer, useRef } from 'react';
 import Autosuggest from 'react-autosuggest';
-import { useSelector } from 'react-redux';
 import withRouter from '@fuse/core/withRouter';
-import { selectFlatNavigation } from 'app/store/fuse/navigationSlice';
+import FuseSvgIcon from '../FuseSvgIcon';
 
 const Root = styled('div')(({ theme }) => ({
   '& .FuseSearch-container': {
@@ -76,12 +74,12 @@ function renderInputComponent(inputProps) {
             variant="outlined"
             {...other}
           />
-          <Icon
+          <FuseSvgIcon
             className="absolute top-0 ltr:right-0 rtl:left-0 h-40 md:h-48 w-48 p-12 pointer-events-none"
             color="action"
           >
-            search
-          </Icon>
+            heroicons-outline:search
+          </FuseSvgIcon>
         </>
       ) : (
         // Standard
@@ -113,7 +111,7 @@ function renderSuggestion(suggestion, { query, isHighlighted }) {
     <MenuItem selected={isHighlighted} component="div">
       <ListItemIcon className="min-w-40">
         {suggestion.icon ? (
-          <Icon>{suggestion.icon}</Icon>
+          <FuseSvgIcon>{suggestion.icon}</FuseSvgIcon>
         ) : (
           <span className="text-20 w-24 font-semibold uppercase text-center">
             {suggestion.title[0]}
@@ -222,7 +220,7 @@ function reducer(state, action) {
 }
 
 function FuseSearch(props) {
-  const navigation = useSelector(selectFlatNavigation);
+  const { navigation } = props;
 
   const [state, dispatch] = useReducer(reducer, initialState);
 
@@ -411,7 +409,7 @@ function FuseSearch(props) {
                     )}
                   />
                   <IconButton onClick={hideSearch} className="mx-8" size="large">
-                    <Icon>close</Icon>
+                    <FuseSvgIcon>heroicons-outline:x</FuseSvgIcon>
                   </IconButton>
                 </div>
               </Paper>
@@ -428,9 +426,10 @@ function FuseSearch(props) {
 
 FuseSearch.propTypes = {};
 FuseSearch.defaultProps = {
+  navigation: [],
   trigger: (
     <IconButton className="w-40 h-40" size="large">
-      <Icon>search</Icon>
+      <FuseSvgIcon>heroicons-outline:search</FuseSvgIcon>
     </IconButton>
   ),
   variant: 'full',
