@@ -1,26 +1,24 @@
 import { memo } from 'react';
+import Box from '@mui/material/Box';
 
 function FuseSplashScreen() {
   return (
     <div id="fuse-splash-screen">
-      <div className="center">
-        <div className="logo">
-          <img width="128" src="assets/images/logos/fuse.svg" alt="logo" />
-        </div>
-        <div className="spinner-wrapper">
-          <div className="spinner">
-            <div className="inner">
-              <div className="gap" />
-              <div className="left">
-                <div className="half-circle" />
-              </div>
-              <div className="right">
-                <div className="half-circle" />
-              </div>
-            </div>
-          </div>
-        </div>
+      <div className="logo">
+        <img width="128" src="assets/images/logo/logo.svg" alt="logo" />
       </div>
+      <Box
+        id="spinner"
+        sx={{
+          '& > div': {
+            backgroundColor: 'palette.secondary.main',
+          },
+        }}
+      >
+        <div className="bounce1" />
+        <div className="bounce2" />
+        <div className="bounce3" />
+      </Box>
     </div>
   );
 }
