@@ -2,20 +2,19 @@ import Typography from '@mui/material/Typography';
 import { useTheme } from '@mui/material/styles';
 import { memo } from 'react';
 import clsx from 'clsx';
-import { useDispatch, useSelector } from 'react-redux';
-import { updateUserSettings } from 'app/auth/store/userSlice';
-import { setDefaultSettings } from 'app/store/fuse/settingsSlice';
 
 function SchemePreview({ theme, className, id, onSelect }) {
   const _theme = useTheme();
   const primaryColor = theme.palette.primary[500]
     ? theme.palette.primary[500]
     : theme.palette.primary.main;
-  const primaryColorContrast = _theme.palette.getContrastText(primaryColor);
+  const primaryColorContrast =
+    theme.palette.primary.contrastText || _theme.palette.getContrastText(primaryColor);
   const secondaryColor = theme.palette.secondary[500]
     ? theme.palette.secondary[500]
     : theme.palette.secondary.main;
-  const secondaryColorContrast = _theme.palette.getContrastText(secondaryColor);
+  const secondaryColorContrast =
+    theme.palette.secondary.contrastText || _theme.palette.getContrastText(secondaryColor);
   const backgroundColor = theme.palette.background.default;
   const backgroundColorContrast = _theme.palette.getContrastText(theme.palette.background.default);
   const paperColor = theme.palette.background.paper;
@@ -31,7 +30,7 @@ function SchemePreview({ theme, className, id, onSelect }) {
           backgroundColor,
           color: backgroundColorContrast,
         }}
-        onClick={() => onSelect(id)}
+        onClick={() => onSelect(theme)}
         type="button"
       >
         <div
@@ -75,28 +74,7 @@ function SchemePreview({ theme, className, id, onSelect }) {
 }
 
 function FuseThemeSchemes(props) {
-  const dispatch = useDispatch();
-  const user = useSelector(({ auth }) => auth.user);
-  const themes = useSelector(({ fuse }) => fuse.settings.themes);
-  const settings = useSelector(({ fuse }) => fuse.settings.current);
-
-  function handleSchemeSelect(themeId) {
-    const newSettings = {
-      ...settings,
-      theme: {
-        main: themeId,
-        navbar: themeId,
-        toolbar: themeId,
-        footer: themeId,
-      },
-    };
-
-    if (user.role === 'guest') {
-      dispatch(setDefaultSettings(newSettings));
-    } else {
-      dispatch(updateUserSettings(newSettings));
-    }
-  }
+  const { themes } = props;
 
   return (
     <div>
@@ -105,7 +83,7 @@ function FuseThemeSchemes(props) {
           .filter(([key, val]) => !(key === 'mainThemeDark' || key === 'mainThemeLight'))
           .map(([key, val]) => (
             <div key={key} className="w-1/2 p-8">
-              <SchemePreview id={key} theme={val} onSelect={handleSchemeSelect} />
+              <SchemePreview id={key} theme={val} onSelect={() => props?.onSelect(val)} />
             </div>
           ))}
       </div>
