@@ -1,8 +1,42 @@
 import { fuseDark } from '@fuse/colors';
-import _ from '@lodash';
 import { lightBlue, red } from '@mui/material/colors';
 import { createTheme } from '@mui/material/styles';
 import qs from 'qs';
+
+const defaultTheme = {
+  palette: {
+    mode: 'light',
+    text: {
+      primary: 'rgb(17, 24, 39)',
+      secondary: 'rgb(107, 114, 128)',
+      disabled: 'rgb(149, 156, 169)',
+    },
+    common: {
+      black: 'rgb(17, 24, 39)',
+      white: 'rgb(255, 255, 255)',
+    },
+    primary: {
+      light: '#bec1c5',
+      main: '#252f3e',
+      dark: '#0d121b',
+      contrastDefaultColor: 'light',
+    },
+    secondary: {
+      light: '#bdf2fa',
+      main: '#22d3ee',
+      dark: '#0cb7e2',
+    },
+    background: {
+      paper: '#FFFFFF',
+      default: '#f6f7f9',
+    },
+    error: {
+      light: '#ffcdd2',
+      main: '#f44336',
+      dark: '#b71c1c',
+    },
+  },
+};
 
 /**
  * SETTINGS DEFAULTS
@@ -11,10 +45,10 @@ export const defaultSettings = {
   customScrollbars: true,
   direction: 'ltr',
   theme: {
-    main: 'default',
-    navbar: 'mainThemeDark',
-    toolbar: 'mainThemeLight',
-    footer: 'mainThemeDark',
+    main: defaultTheme,
+    navbar: defaultTheme,
+    toolbar: defaultTheme,
+    footer: defaultTheme,
   },
 };
 
@@ -44,6 +78,11 @@ export const defaultThemeOptions = {
     fontWeightMedium: 500,
   },
   components: {
+    MuiDateTimePicker: {
+      defaultProps: {
+        PopperProps: { className: 'z-9999' },
+      },
+    },
     MuiAppBar: {
       defaultProps: {
         enableColorOnDark: true,
@@ -54,17 +93,32 @@ export const defaultThemeOptions = {
         },
       },
     },
+    MuiButtonBase: {
+      defaultProps: {
+        disableRipple: true,
+      },
+    },
     MuiButton: {
+      defaultProps: {
+        variant: 'text',
+        color: 'inherit',
+      },
       styleOverrides: {
         root: {
           textTransform: 'none',
-          borderRadius: '18px',
+          // lineHeight: 1,
+        },
+        sizeMedium: {
+          borderRadius: 20,
+          height: 40,
+          minHeight: 40,
+          maxHeight: 40,
         },
         sizeSmall: {
           borderRadius: '15px',
         },
         sizeLarge: {
-          borderRadius: '21px',
+          borderRadius: '28px',
         },
         contained: {
           boxShadow: 'none',
@@ -75,6 +129,9 @@ export const defaultThemeOptions = {
       },
     },
     MuiButtonGroup: {
+      defaultProps: {
+        color: 'secondary',
+      },
       styleOverrides: {
         contained: {
           borderRadius: 18,
@@ -97,6 +154,9 @@ export const defaultThemeOptions = {
     },
     MuiPaper: {
       styleOverrides: {
+        root: {
+          backgroundImage: 'none',
+        },
         rounded: {
           borderRadius: 16,
         },
@@ -106,6 +166,34 @@ export const defaultThemeOptions = {
       styleOverrides: {
         paper: {
           borderRadius: 8,
+        },
+      },
+    },
+    MuiTextField: {
+      defaultProps: {
+        color: 'secondary',
+      },
+    },
+    MuiInputLabel: {
+      defaultProps: {
+        color: 'secondary',
+      },
+    },
+    MuiSelect: {
+      defaultProps: {
+        color: 'secondary',
+      },
+    },
+    MuiOutlinedInput: {
+      defaultProps: {
+        color: 'secondary',
+      },
+    },
+    MuiInputBase: {
+      styleOverrides: {
+        root: {
+          minHeight: 40,
+          lineHeight: 1,
         },
       },
     },
@@ -139,18 +227,28 @@ export const defaultThemeOptions = {
         color: 'secondary',
       },
     },
+    MuiTypography: {
+      variants: [
+        {
+          props: { color: 'text.secondary' },
+          style: {
+            color: 'text.secondary',
+          },
+        },
+      ],
+    },
   },
 };
 
 export const mustHaveThemeOptions = {
   typography: {
     htmlFontSize: 10,
-    fontSize: 13,
+    fontSize: 14,
     body1: {
-      fontSize: '1.3rem',
+      fontSize: '1.4rem',
     },
     body2: {
-      fontSize: '1.3rem',
+      fontSize: '1.4rem',
     },
   },
 };
@@ -215,39 +313,6 @@ export function extendThemeWithMixins(obj) {
       borderBottomWidth: width,
       borderStyle: 'solid',
       borderColor: theme.palette.divider,
-    }),
-  };
-}
-
-export function mainThemeVariations(theme) {
-  return {
-    mainThemeDark: _.merge({}, theme, {
-      palette: {
-        mode: 'dark',
-        background: {
-          paper: '#1E2125',
-          default: '#121212',
-        },
-        text: {
-          primary: 'rgb(255,255,255)',
-          secondary: 'rgb(229, 231, 235)',
-          disabled: 'rgb(156, 163, 175)',
-        },
-      },
-    }),
-    mainThemeLight: _.merge({}, theme, {
-      palette: {
-        mode: 'light',
-        background: {
-          paper: '#FFFFFF',
-          default: '#F7F7F7',
-        },
-        text: {
-          primary: 'rgb(17, 24, 39)',
-          secondary: 'rgb(107, 114, 128)',
-          disabled: 'rgb(149, 156, 169)',
-        },
-      },
     }),
   };
 }
