@@ -2,7 +2,7 @@
  * Authorization Roles
  */
 const authRoles = {
-  admin: ['admin'],
+  admin: ['admin', 'Authenticated'],
   staff: ['admin', 'staff'],
   user: ['admin', 'staff', 'user'],
   onlyGuest: [],
