@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as authRoles } from './authRoles';
-export { default as Auth } from './Auth';
