@@ -1,20 +1,233 @@
-import Icon from '@mui/material/Icon';
-import Input from '@mui/material/Input';
-import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
 import { motion } from 'framer-motion';
 import { Link } from 'react-router-dom';
+import Box from '@mui/material/Box';
 
 function Error404Page() {
   return (
     <div className="flex flex-col flex-1 items-center justify-center p-16">
-      <div className="max-w-512 text-center">
+      <div className="w-full max-w-3xl text-center">
         <motion.div
           initial={{ opacity: 0, scale: 0.6 }}
           animate={{ opacity: 1, scale: 1, transition: { delay: 0.1 } }}
         >
-          <Typography variant="h1" color="inherit" className="font-medium mb-16">
-            404
+          <Box
+            component="svg"
+            width="100%"
+            height="100%"
+            viewBox="0 0 1075 585"
+            fill="none"
+            preserveAspectRatio="xMidYMax slice"
+            xmlns="http://www.w3.org/2000/svg"
+            sx={{ color: 'secondary.main' }}
+          >
+            <g clipPath="url(#clip0)">
+              <path
+                d="M520.426 167.01C434.482 167.01 372.775 222.149 372.775 350.808C372.775 496.621 434.482 535.218 520.426 535.218C606.37 535.218 671.753 492.945 671.753 350.808C671.753 198.868 606.37 167.01 520.426 167.01ZM520.991 486.818C461.464 486.818 430.365 451.895 430.365 350.902C430.365 261.79 462.737 214.797 522.264 214.797C581.791 214.797 614.163 245.665 614.163 350.902C614.163 449.349 580.517 486.818 520.991 486.818Z"
+                className="text-gray-800 dark:text-gray-600"
+                fill="currentColor"
+              />
+              <path
+                d="M321.311 433.517H285.777V328.877C285.777 323.32 283.569 317.989 279.639 314.059C275.709 310.129 270.379 307.922 264.821 307.922H256.495C253.743 307.922 251.018 308.464 248.475 309.517C245.933 310.57 243.623 312.113 241.677 314.059C239.731 316.005 238.187 318.315 237.134 320.858C236.081 323.4 235.539 326.125 235.539 328.877V433.517H135.309C133.247 433.517 131.22 432.984 129.425 431.97C127.629 430.955 126.127 429.494 125.062 427.728C123.998 425.962 123.408 423.951 123.35 421.89C123.292 419.829 123.768 417.788 124.731 415.965L230.814 215.184C232.136 212.681 232.938 209.936 233.17 207.114C233.402 204.293 233.061 201.453 232.165 198.768C231.27 196.082 229.84 193.605 227.961 191.487C226.082 189.37 223.793 187.654 221.233 186.445L214.971 183.488C210.108 181.192 204.549 180.853 199.444 182.541C194.338 184.23 190.077 187.816 187.542 192.558L58.1602 434.591C55.957 438.712 54.8043 443.314 54.8043 447.987V447.987C54.8043 451.719 55.5393 455.414 56.9673 458.861C58.3954 462.309 60.4885 465.441 63.1271 468.08C65.7658 470.719 68.8983 472.812 72.3459 474.24C75.7935 475.668 79.4885 476.403 83.2202 476.403H235.539V542.57C235.539 545.869 236.189 549.135 237.451 552.183C238.713 555.23 240.564 557.999 242.896 560.332C245.229 562.664 247.998 564.515 251.045 565.777C254.093 567.039 257.359 567.689 260.658 567.689H260.658C263.957 567.689 267.223 567.039 270.271 565.777C273.318 564.515 276.087 562.664 278.42 560.332C280.752 557.999 282.603 555.23 283.865 552.183C285.127 549.135 285.777 545.869 285.777 542.57V476.403H321.311C326.998 476.403 332.452 474.144 336.474 470.122C340.495 466.101 342.754 460.647 342.754 454.96V454.96C342.754 449.273 340.495 443.819 336.474 439.797C332.453 435.776 326.998 433.517 321.311 433.517V433.517Z"
+                fill="currentColor"
+              />
+              <path
+                d="M979.308 433.517H943.774V328.877C943.774 323.32 941.566 317.989 937.636 314.059C933.706 310.129 928.376 307.922 922.818 307.922H914.491C911.739 307.922 909.014 308.464 906.472 309.517C903.929 310.57 901.619 312.113 899.673 314.059C897.727 316.005 896.184 318.315 895.131 320.858C894.077 323.4 893.535 326.125 893.535 328.877V433.517H793.305C791.243 433.517 789.216 432.984 787.421 431.97C785.626 430.955 784.123 429.494 783.059 427.728C781.995 425.962 781.405 423.951 781.347 421.89C781.289 419.829 781.764 417.788 782.728 415.965L888.81 215.184C890.133 212.681 890.934 209.936 891.167 207.114C891.399 204.293 891.057 201.453 890.162 198.768C889.266 196.082 887.836 193.605 885.957 191.487C884.078 189.37 881.79 187.654 879.23 186.445L872.967 183.488C868.105 181.192 862.546 180.853 857.44 182.541C852.334 184.23 848.073 187.816 845.538 192.558L716.157 434.591C713.953 438.712 712.801 443.314 712.801 447.987V447.987C712.801 455.523 715.795 462.751 721.124 468.08C726.453 473.409 733.68 476.403 741.217 476.403H893.535V542.57C893.535 549.232 896.182 555.621 900.893 560.332C905.603 565.043 911.992 567.689 918.654 567.689V567.689C925.316 567.689 931.706 565.043 936.416 560.332C941.127 555.621 943.773 549.232 943.773 542.57V476.403H979.308C984.995 476.403 990.449 474.144 994.47 470.122C998.492 466.101 1000.75 460.647 1000.75 454.96V454.96C1000.75 449.273 998.492 443.819 994.47 439.797C990.449 435.776 984.995 433.517 979.308 433.517Z"
+                fill="currentColor"
+              />
+              <path
+                d="M471.141 31.1516C478.266 31.1516 484.042 25.3756 484.042 18.2505C484.042 11.1253 478.266 5.34927 471.141 5.34927C464.016 5.34927 458.24 11.1253 458.24 18.2505C458.24 25.3756 464.016 31.1516 471.141 31.1516Z"
+                className="text-gray-800"
+                fill="currentColor"
+              />
+              <path
+                d="M502.607 93.7695C522.592 93.7695 538.794 72.7785 538.794 46.8848C538.794 20.991 522.592 0 502.607 0C482.622 0 466.421 20.991 466.421 46.8848C466.421 72.7785 482.622 93.7695 502.607 93.7695Z"
+                className="text-gray-800"
+                fill="currentColor"
+              />
+              <path
+                opacity="0.1"
+                d="M502.607 79.6097C484.48 79.6097 469.466 62.3392 466.832 39.8048C466.558 42.1548 466.421 44.5188 466.421 46.8848C466.421 72.7785 482.622 93.7695 502.607 93.7695C522.592 93.7695 538.794 72.7785 538.794 46.8848C538.794 44.5188 538.657 42.1548 538.383 39.8048C535.749 62.3392 520.735 79.6097 502.607 79.6097Z"
+                fill="black"
+              />
+              <path
+                opacity="0.1"
+                d="M576.242 184.189C575.465 187.382 572.114 189.94 568.403 191.727C560.602 195.483 550.978 196.605 541.644 196.981C539.262 197.126 536.873 197.081 534.499 196.849C529.898 196.404 525.582 194.42 522.249 191.217C522.011 191.062 521.816 190.849 521.683 190.599C521.549 190.348 521.481 190.067 521.486 189.783C521.49 189.499 521.566 189.221 521.707 188.975C521.849 188.728 522.05 188.521 522.292 188.374C527.262 182.915 535.555 179.567 543.909 177.157C550.243 175.33 561.29 170.597 568.199 171.235C574.89 171.854 577.223 180.157 576.242 184.189Z"
+                fill="black"
+              />
+              <path
+                d="M477.044 160.321C474.662 164.989 474.231 170.411 475.845 175.396C477.487 180.344 480.56 184.693 484.674 187.893C494.769 195.985 508.662 197.098 521.599 196.986C532.937 196.887 544.259 196.111 555.504 194.662C559.273 194.328 562.976 193.468 566.506 192.108C570.728 190.281 574.439 186.93 575.942 182.582C577.445 178.234 576.378 172.909 572.808 170.008C569.671 167.458 565.345 167.107 561.31 166.865L538.223 165.477C540.504 163.169 543.541 161.863 546.318 160.185C550.004 157.959 552.452 154.22 555.132 150.85C559.26 145.916 563.752 141.297 568.57 137.033C569.322 141.344 572.048 144.883 575.258 147.857C578.467 150.831 582.166 153.226 585.501 156.057C593.58 162.916 599.397 172.224 607.894 178.558C613.738 177.022 619.202 174.294 623.942 170.546L620.64 164.564C620.365 163.949 619.936 163.414 619.396 163.011C618.461 162.436 617.256 162.733 616.17 162.57C613.27 162.135 612.003 158.786 610.865 156.083C607.74 148.661 601.757 142.903 596.647 136.679C592.875 132.084 589.529 127.157 586.651 121.956C584.934 118.854 583.363 115.625 581.033 112.953C578.704 110.281 575.44 108.18 571.895 108.172C568.763 108.166 565.874 109.756 563.181 111.355C554.978 116.227 547.087 121.604 539.551 127.454C535.607 130.516 531.688 133.751 527.061 135.628C525.114 136.419 523.061 136.959 521.192 137.92C517.919 139.605 515.435 142.455 512.761 144.985C503.223 154.008 490.165 159.854 477.044 160.321Z"
+                className="text-gray-700"
+                fill="currentColor"
+              />
+              <path
+                d="M525.32 96.098C525.213 98.8561 524.626 101.575 523.586 104.131C523.059 105.407 522.275 106.56 521.282 107.518C520.29 108.477 519.111 109.221 517.818 109.704C515.983 110.196 514.059 110.258 512.195 109.887C510.57 109.717 509.004 109.183 507.614 108.326C505.449 106.837 504.369 104.182 503.976 101.583C503.583 98.9866 503.759 96.334 503.507 93.7191C503.262 91.6071 502.86 89.5162 502.305 87.4637C502.135 86.7808 501.962 86.1012 501.792 85.4215C501.626 84.7418 501.462 84.0622 501.311 83.3825C501.172 82.7437 501.043 82.1018 500.933 81.4568C500.842 80.9124 500.76 80.3649 500.7 79.8142C500.645 78.1817 500.373 76.5638 499.892 75.003C499.44 73.9908 498.874 73.0333 498.205 72.149C497.919 71.74 497.642 71.3309 497.387 70.9061C496.65 69.8022 496.307 68.4816 496.415 67.1585C496.445 66.9539 496.495 66.7527 496.563 66.5575C496.588 66.482 496.616 66.4065 496.648 66.3341C496.931 66.2572 497.191 66.1124 497.405 65.9123C497.62 65.7121 497.782 65.4626 497.878 65.1855C497.986 64.5902 497.935 63.9769 497.73 63.4077C497.495 62.6123 497.173 61.8451 496.77 61.12C496.506 60.6418 495.609 59.7796 495.603 59.232C495.59 58.1307 497.743 57.2434 498.614 56.8091C501.275 55.4769 504.242 54.8753 507.211 55.0659C509.115 55.1949 511.179 55.7802 512.23 57.3755C513.174 58.8135 513.048 60.6763 513.451 62.3472C513.493 62.5901 513.596 62.8184 513.75 63.0111C513.955 63.1983 514.213 63.317 514.489 63.351C514.955 63.4437 515.433 63.4628 515.905 63.4076H515.908C516.033 63.3977 516.157 63.3799 516.28 63.3541C517.015 63.2185 517.71 62.9184 518.312 62.4762C518.702 62.3724 519.096 62.2874 519.492 62.2213C519.508 62.2197 519.524 62.2165 519.539 62.2119C521.116 67.1867 521.32 72.5202 522.611 77.5863C523.249 80.1068 524.165 82.5958 524.294 85.1886C524.303 85.3018 524.307 85.4183 524.31 85.5347C524.329 86.3686 524.263 87.2087 524.297 88.0394C524.309 88.4135 524.351 88.786 524.423 89.1533C524.564 89.7441 524.74 90.3262 524.949 90.8966L524.958 90.9249C525.018 91.1074 525.074 91.2931 525.122 91.4787C525.156 91.614 525.188 91.7462 525.213 91.8815L525.216 91.8909C525.449 93.2813 525.484 94.6978 525.32 96.098V96.098Z"
+                fill="#FBBEBE"
+              />
+              <path
+                d="M503.237 63.5619C513.49 63.5619 521.802 55.25 521.802 44.9968C521.802 34.7436 513.49 26.4317 503.237 26.4317C492.983 26.4317 484.672 34.7436 484.672 44.9968C484.672 55.25 492.983 63.5619 503.237 63.5619Z"
+                fill="#FBBEBE"
+              />
+              <path
+                d="M621.103 146.148C620.565 141.083 618.765 136.234 615.867 132.045C616.349 134.909 616.831 137.771 617.315 140.632C616.714 140.806 616.075 140.801 615.477 140.617C614.879 140.433 614.347 140.079 613.948 139.597C613.166 138.621 612.568 137.511 612.182 136.321C610.073 131.085 607.315 126.135 603.973 121.586C592.154 120.601 580.471 115.702 570.077 109.99C559.681 104.282 549.952 97.4385 539.694 91.4882C537.915 90.4132 536.062 89.4668 534.149 88.6562C530.94 87.3692 527.337 86.7022 524.294 85.1886C524.089 85.0879 523.885 84.981 523.683 84.8708C522.996 84.497 522.345 84.0598 521.739 83.565C521.411 83.2989 521.1 83.0131 520.807 82.7091C519.289 80.9429 518.086 78.9292 517.249 76.7557C515.616 73.1434 515.033 68.7853 513.404 65.1729C514.153 64.4728 514.994 63.8789 515.905 63.4077H515.908C516.04 63.3385 516.172 63.2724 516.308 63.2094C516.953 62.9057 517.624 62.6603 518.312 62.4763C518.702 62.3724 519.096 62.2875 519.492 62.2214C519.508 62.2198 519.524 62.2166 519.539 62.212C521.378 61.9124 523.262 62.0614 525.03 62.6462C527.617 63.521 529.797 65.2673 531.972 66.9193C535.504 69.4079 538.892 72.0952 542.12 74.9684C543.913 76.6864 545.549 78.5744 547.478 80.1352C549.081 81.3722 550.771 82.4922 552.535 83.4863C562.972 89.704 573.411 95.9207 583.85 102.136C587.384 104.241 590.921 106.347 594.376 108.574C599.624 111.96 604.684 115.629 609.741 119.295C612.721 121.454 615.699 123.613 618.677 125.774C619.654 126.397 620.512 127.19 621.21 128.115C621.827 129.165 622.21 130.336 622.333 131.548C623.081 136.444 622.659 141.447 621.103 146.148V146.148Z"
+                fill="#FBBEBE"
+              />
+              <path
+                d="M529.923 110.06C528.671 115.676 522.994 119.072 520.027 124.005C516.339 130.129 517.843 137.246 519.228 144.26C516.698 146.441 514.167 148.622 511.635 150.802C510.238 152.001 508.841 153.206 507.337 154.267C505.144 155.736 502.852 157.052 500.477 158.206C498.237 159.358 495.996 160.511 493.756 161.664C491.673 162.835 489.466 163.77 487.176 164.452C482.762 165.585 478.073 164.792 473.706 163.48C473.191 163.345 472.695 163.143 472.233 162.879C471.961 162.716 471.723 162.501 471.534 162.246C471.104 161.474 470.947 160.579 471.091 159.707C471.534 152.357 473.331 145.16 475.122 138.017C476.079 133.732 477.365 129.527 478.97 125.44C480.544 121.938 482.353 118.547 484.385 115.289C486.628 111.191 489.285 107.333 492.315 103.776C493.79 102.407 495.031 100.806 495.99 99.0369C497.107 96.4882 496.676 93.5649 496.903 90.7896C497.076 88.7045 497.629 86.6686 498.536 84.7827V84.7796C499.02 83.7638 499.604 82.7983 500.279 81.8973C500.584 82.6084 500.889 83.3185 501.195 84.0275C501.396 84.4932 501.595 84.9579 501.792 85.4215C502.05 86.0194 502.306 86.6152 502.56 87.2088C503.293 88.9143 504.025 90.6187 504.757 92.3221C505.547 94.4675 506.672 96.4739 508.092 98.2661C508.741 99.1838 509.617 99.9176 510.634 100.396C511.651 100.874 512.775 101.08 513.895 100.994C515.016 100.908 516.095 100.533 517.028 99.9055C517.96 99.2779 518.714 98.4192 519.215 97.4133C520.273 95.5357 521.039 93.5081 521.487 91.4001C522.109 88.8335 522.194 86.1663 521.739 83.565C520.981 78.5556 518.85 73.4392 517.302 68.7853C516.672 67.2336 516.307 65.587 516.223 63.9143C516.226 63.7264 516.245 63.539 516.279 63.3542C516.289 63.307 516.298 63.2567 516.308 63.2095C516.953 62.9058 517.624 62.6604 518.312 62.4763C518.702 62.3725 519.096 62.2875 519.492 62.2214C520.855 69.5971 522.219 76.9728 523.586 84.3485C523.617 84.5216 523.652 84.6947 523.684 84.8709C523.879 85.9281 524.077 86.9854 524.297 88.0395C524.495 88.9961 524.709 89.9495 524.948 90.8967L524.958 90.925C525.04 91.246 525.125 91.5638 525.213 91.8816L525.216 91.891C525.458 92.7595 525.729 93.6216 526.034 94.4712C527.884 99.5908 531.11 104.745 529.923 110.06Z"
+                fill="currentColor"
+              />
+              <path
+                d="M605.267 188.993C605.209 190.075 605.419 191.154 605.877 192.136C607.039 194.259 609.827 194.742 612.24 194.941C626.332 196.099 640.51 195.667 654.505 193.653C655.265 193.601 656.001 193.372 656.657 192.985C658.647 191.59 657.42 188.19 655.247 187.102C653.074 186.013 650.496 186.276 648.112 185.804C645.743 185.208 643.488 184.23 641.434 182.908C636.599 180.13 631.835 177.234 627.141 174.22C624.346 172.425 621.048 170.51 617.901 171.574C609.376 174.458 605.211 179.796 605.267 188.993Z"
+                className="text-gray-700"
+                fill="currentColor"
+              />
+              <path
+                d="M501.383 82.9892V82.9923C501.361 83.1245 501.339 83.2535 501.311 83.3825C501.276 83.5965 501.235 83.8136 501.195 84.0275C500.804 86.0256 500.326 88.0143 499.775 89.9904C498.265 95.4498 496.242 100.821 494.234 106.114C488.847 120.305 483.459 134.497 478.07 148.691C476.711 152.265 475.354 155.843 473.809 159.345C473.293 160.513 472.762 161.689 472.233 162.879C470.339 167.136 468.52 171.532 467.916 176.063C467.353 179.045 467.288 182.1 467.724 185.103C468.328 188.074 470.131 190.975 472.954 192.073C474.332 192.611 476.305 193.332 475.927 194.764C475.723 195.55 474.829 195.934 474.027 196.066C472.967 196.235 471.883 196.169 470.852 195.872C469.821 195.575 468.868 195.055 468.061 194.348C466.553 193.011 465.395 190.931 463.391 190.708C462.003 190.557 460.726 191.387 459.454 191.969C454.269 194.339 448.227 192.681 442.781 190.997C443.208 188.791 445.424 187.338 447.623 186.897C449.823 186.453 452.098 186.768 454.341 186.708C455.06 186.747 455.773 186.563 456.383 186.183C456.931 185.68 457.315 185.025 457.485 184.301C458.93 180.019 460.005 175.621 460.697 171.155C461.421 166.488 461.723 161.749 462.755 157.139C463.558 153.543 464.797 150.06 465.861 146.529C467.918 139.26 469.606 131.892 470.921 124.452C471.585 121.362 472.061 118.233 472.346 115.085C472.526 112.473 472.441 109.849 472.692 107.246C473.115 103.861 473.821 100.517 474.804 97.2497C475.767 93.6845 476.728 90.1215 477.689 86.5606C479.115 81.29 480.644 75.8149 484.256 71.7179C485.511 70.2756 487.048 69.1064 488.773 68.2826C490.498 67.4588 492.374 66.998 494.285 66.9288C495.001 66.9279 495.715 67.0049 496.415 67.1585C496.531 67.1868 496.648 67.2151 496.764 67.2466C497.526 67.4459 498.258 67.7487 498.938 68.1465C500.052 67.4385 500.619 67.6871 501.358 68.779C502.006 69.9345 502.312 71.2509 502.239 72.5738C502.302 76.0655 502.015 79.5546 501.383 82.9892Z"
+                fill="#FBBEBE"
+              />
+              <path
+                d="M502.607 87.1616L502.56 87.2088L502.305 87.4637L499.775 89.9904L498.536 84.7827L494.285 66.9287C495.001 66.9279 495.715 67.0049 496.415 67.1584C496.445 66.9539 496.495 66.7527 496.563 66.5574L496.764 67.2465L498.205 72.149L500.933 81.4567L501.383 82.9891V82.9923L502.607 87.1616Z"
+                fill="currentColor"
+              />
+              <path
+                d="M518.661 30.2028C518.659 31.856 518.331 33.4926 517.696 35.019C517.061 36.5454 516.132 37.9317 514.961 39.0986C514.537 39.5234 514.082 39.9165 513.601 40.275C511.429 41.9119 508.784 42.7979 506.065 42.7991H490.969C487.633 42.794 484.436 41.4652 482.079 39.1047C479.722 36.7441 478.398 33.5447 478.398 30.2089C478.398 26.8732 479.722 23.6737 482.079 21.3132C484.436 18.9526 487.633 17.6239 490.969 17.6187H506.065C509.401 17.6274 512.598 18.9556 514.958 21.3134C517.318 23.6711 518.649 26.8668 518.661 30.2028V30.2028Z"
+                className="text-gray-800"
+                fill="currentColor"
+              />
+              <path
+                d="M532.815 31.1516C539.94 31.1516 545.716 25.3756 545.716 18.2505C545.716 11.1253 539.94 5.34927 532.815 5.34927C525.69 5.34927 519.914 11.1253 519.914 18.2505C519.914 25.3756 525.69 31.1516 532.815 31.1516Z"
+                className="text-gray-800"
+                fill="currentColor"
+              />
+              <path
+                opacity="0.1"
+                d="M532.5 5.34927C532.342 5.34927 532.185 5.35556 532.028 5.36121C535.366 5.48254 538.526 6.89362 540.844 9.29756C543.162 11.7015 544.457 14.9109 544.457 18.2505C544.457 21.59 543.162 24.7994 540.844 27.2033C538.526 29.6073 535.366 31.0184 532.028 31.1397C532.185 31.1453 532.342 31.1516 532.5 31.1516C535.922 31.1516 539.203 29.7924 541.623 27.373C544.042 24.9535 545.401 21.6721 545.401 18.2505C545.401 14.8289 544.042 11.5474 541.623 9.12794C539.203 6.70851 535.922 5.34927 532.5 5.34927V5.34927Z"
+                fill="black"
+              />
+              <path
+                opacity="0.1"
+                d="M471.141 5.34927C471.299 5.34927 471.456 5.35556 471.613 5.36121C468.276 5.48254 465.116 6.89362 462.797 9.29756C460.479 11.7015 459.184 14.9109 459.184 18.2505C459.184 21.59 460.479 24.7994 462.797 27.2033C465.116 29.6073 468.276 31.0184 471.613 31.1397C471.456 31.1453 471.299 31.1516 471.141 31.1516C467.721 31.1486 464.443 29.7881 462.026 27.3689C459.609 24.9498 458.251 21.6701 458.251 18.2505C458.251 14.8308 459.609 11.5511 462.026 9.13196C464.443 6.71284 467.721 5.35228 471.141 5.34927Z"
+                fill="black"
+              />
+              <path
+                opacity="0.1"
+                d="M513.601 40.275C511.429 41.9119 508.784 42.7979 506.065 42.7991H490.969C488.25 42.7978 485.605 41.9119 483.433 40.275H513.601Z"
+                fill="black"
+              />
+              <path
+                d="M611.086 133.756C611.086 133.756 614.338 143.128 617.315 140.632L614.942 128.382L611.086 133.756Z"
+                fill="#FBBEBE"
+              />
+              <path
+                d="M1006.86 419.55C1009.89 420.651 1013.14 421.034 1016.34 420.667C1019.54 420.301 1022.62 419.196 1025.32 417.439C1028.02 415.683 1030.28 413.322 1031.91 410.544C1033.55 407.766 1034.51 404.647 1034.74 401.431C1034.96 398.216 1034.43 394.992 1033.2 392.015C1031.96 389.038 1030.05 386.388 1027.62 384.275C1025.19 382.161 1022.3 380.642 1019.17 379.837C1016.05 379.032 1012.79 378.962 1009.64 379.634L1010.16 390.347L1005.09 381.166C1000.11 383.503 996.271 387.72 994.404 392.889C993.741 394.699 993.338 396.595 993.207 398.519C992.887 403.032 994.048 407.527 996.512 411.322C998.976 415.117 1002.61 418.006 1006.86 419.55V419.55Z"
+                fill="#57B894"
+              />
+              <path
+                d="M1031.94 543.793C1030.15 534.683 1037.9 526.606 1045.56 521.357C1053.22 516.107 1062.17 510.949 1064.78 502.04C1068.53 489.236 1057.35 477.51 1048.64 467.398C1042.19 459.894 1036.64 451.65 1032.12 442.841C1030.31 439.308 1028.64 435.615 1028.17 431.674C1027.48 425.999 1029.31 420.351 1031.14 414.938C1037.26 396.903 1043.8 379.02 1050.77 361.288"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M1007.72 416.792C1010.75 417.893 1013.99 418.276 1017.2 417.909C1020.4 417.543 1023.47 416.438 1026.17 414.682C1028.88 412.925 1031.13 410.565 1032.77 407.787C1034.4 405.009 1035.37 401.889 1035.59 398.673C1035.82 395.458 1035.29 392.234 1034.06 389.257C1032.82 386.28 1030.91 383.63 1028.48 381.517C1026.04 379.403 1023.15 377.884 1020.03 377.079C1016.91 376.274 1013.64 376.204 1010.49 376.876L1011.01 387.589L1005.95 378.408C1000.97 380.745 997.128 384.962 995.262 390.131C994.599 391.941 994.196 393.837 994.064 395.761C993.745 400.274 994.905 404.769 997.369 408.564C999.833 412.359 1003.47 415.248 1007.72 416.792V416.792Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M1029.4 357.588C1030.59 354.308 1032.58 351.38 1035.2 349.073C1037.82 346.767 1040.97 345.157 1044.38 344.391L1046.09 354.574L1049.27 343.879C1053.68 343.944 1057.96 345.41 1061.48 348.064C1065 350.718 1067.59 354.423 1068.87 358.645C1070.15 362.866 1070.06 367.386 1068.6 371.551C1067.15 375.715 1064.41 379.31 1060.78 381.815C1057.15 384.32 1052.81 385.607 1048.4 385.489C1043.99 385.371 1039.73 383.854 1036.24 381.158C1032.75 378.462 1030.21 374.726 1028.98 370.49C1027.74 366.253 1027.89 361.735 1029.4 357.588H1029.4Z"
+                fill="#57B894"
+              />
+              <path
+                d="M1030.54 353.911C1031.73 350.631 1033.72 347.703 1036.34 345.396C1038.96 343.09 1042.12 341.479 1045.52 340.713L1047.23 350.897L1050.41 340.201C1054.82 340.267 1059.1 341.732 1062.62 344.387C1066.15 347.041 1068.74 350.746 1070.02 354.968C1071.3 359.189 1071.2 363.709 1069.75 367.873C1068.29 372.038 1065.55 375.633 1061.92 378.138C1058.29 380.643 1053.96 381.93 1049.55 381.812C1045.14 381.694 1040.88 380.177 1037.38 377.481C1033.89 374.785 1031.35 371.049 1030.12 366.813C1028.89 362.576 1029.04 358.058 1030.54 353.911H1030.54Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M1044.99 467.604C1049.8 469.347 1055.07 469.261 1059.82 467.363C1064.56 465.464 1068.44 461.886 1070.72 457.311C1072.99 452.735 1073.51 447.482 1072.16 442.553C1070.81 437.623 1067.7 433.362 1063.41 430.582L1060.97 438.801L1059.24 428.497C1059.22 428.489 1059.2 428.482 1059.18 428.477C1056.62 427.543 1053.89 427.125 1051.16 427.246C1048.42 427.367 1045.74 428.025 1043.27 429.182C1040.79 430.339 1038.56 431.973 1036.72 433.991C1034.87 436.008 1033.44 438.369 1032.51 440.94C1031.58 443.51 1031.16 446.238 1031.29 448.97C1031.41 451.701 1032.07 454.381 1033.23 456.857C1034.39 459.334 1036.02 461.557 1038.04 463.401C1040.06 465.245 1042.42 466.674 1044.99 467.604H1044.99Z"
+                fill="#57B894"
+              />
+              <path
+                d="M1045.98 463.88C1050.79 465.622 1056.06 465.536 1060.81 463.638C1065.55 461.74 1069.43 458.162 1071.71 453.586C1073.99 449.011 1074.5 443.758 1073.15 438.828C1071.8 433.899 1068.69 429.638 1064.4 426.857L1061.96 435.076L1060.23 424.772C1060.21 424.764 1060.19 424.758 1060.17 424.752C1057.61 423.818 1054.88 423.4 1052.15 423.521C1049.41 423.642 1046.73 424.3 1044.26 425.457C1041.78 426.615 1039.55 428.249 1037.71 430.266C1035.86 432.284 1034.43 434.645 1033.5 437.215C1032.57 439.785 1032.15 442.514 1032.28 445.245C1032.4 447.976 1033.06 450.656 1034.22 453.133C1035.38 455.609 1037.01 457.833 1039.03 459.677C1041.05 461.521 1043.41 462.949 1045.98 463.88H1045.98Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M1023.32 502.167C1024.15 506.867 1026.57 511.14 1030.17 514.27C1033.77 517.4 1038.34 519.195 1043.11 519.356C1047.88 519.517 1052.56 518.032 1056.37 515.152C1060.17 512.272 1062.87 508.171 1064.01 503.537C1065.15 498.903 1064.66 494.018 1062.63 489.701C1060.59 485.384 1057.14 481.897 1052.84 479.826C1048.54 477.755 1043.66 477.225 1039.02 478.326C1034.37 479.426 1030.25 482.09 1027.33 485.87L1036.61 499.826L1023.94 492.418C1022.95 495.572 1022.74 498.915 1023.32 502.167V502.167Z"
+                fill="#57B894"
+              />
+              <path
+                d="M1024.18 499.409C1025.01 504.109 1027.42 508.382 1031.03 511.512C1034.63 514.642 1039.2 516.437 1043.97 516.598C1048.74 516.759 1053.42 515.274 1057.22 512.394C1061.03 509.514 1063.73 505.413 1064.87 500.779C1066.01 496.145 1065.52 491.26 1063.49 486.943C1061.45 482.626 1058 479.139 1053.7 477.068C1049.4 474.997 1044.52 474.467 1039.87 475.568C1035.23 476.668 1031.1 479.332 1028.19 483.112L1037.46 497.068L1024.79 489.661C1023.81 492.814 1023.6 496.157 1024.18 499.409V499.409Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M9.00746 470.249C10.9978 470.972 13.1289 471.224 15.2329 470.983C17.337 470.743 19.3563 470.017 21.1319 468.862C22.9075 467.708 24.3908 466.157 25.4648 464.332C26.5388 462.507 27.1742 460.457 27.3208 458.345C27.4675 456.232 27.1215 454.114 26.3101 452.158C25.4987 450.202 24.244 448.461 22.645 447.073C21.0459 445.684 19.1463 444.686 17.0957 444.157C15.0451 443.628 12.8997 443.582 10.8285 444.024L11.1707 451.062L7.84085 445.03C4.57278 446.566 2.04805 449.336 0.821494 452.732C0.385892 453.922 0.121078 455.167 0.034953 456.431C-0.17523 459.397 0.586979 462.35 2.20607 464.843C3.82517 467.337 6.2129 469.234 9.00746 470.249V470.249Z"
+                fill="#57B894"
+              />
+              <path
+                d="M25.4837 551.879C24.3083 545.893 29.4017 540.587 34.4329 537.138C39.464 533.689 45.3427 530.3 47.0593 524.446C49.5264 516.034 42.1776 508.33 36.4579 501.686C32.2144 496.756 28.5702 491.34 25.6016 485.551C24.311 483.278 23.4319 480.794 23.0049 478.215C22.5539 474.486 23.7519 470.775 24.9588 467.219C28.9794 455.37 33.2782 443.62 37.8551 431.97"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M9.57064 468.437C11.5609 469.16 13.6921 469.412 15.7961 469.171C17.9001 468.931 19.9195 468.205 21.6951 467.05C23.4707 465.896 24.9539 464.345 26.0279 462.52C27.1019 460.695 27.7373 458.645 27.884 456.533C28.0307 454.42 27.6847 452.302 26.8733 450.346C26.0618 448.39 24.8072 446.649 23.2081 445.261C21.6091 443.872 19.7094 442.874 17.6589 442.345C15.6083 441.816 13.4628 441.77 11.3916 442.212L11.7339 449.25L8.40403 443.218C5.13595 444.754 2.61122 447.524 1.38467 450.92C0.949069 452.11 0.684259 453.355 0.598129 454.619C0.387932 457.585 1.15014 460.538 2.76923 463.031C4.38833 465.525 6.77607 467.422 9.57064 468.437V468.437Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M23.8124 429.539C24.5935 427.384 25.9044 425.46 27.6244 423.945C29.3444 422.429 31.4181 421.371 33.6545 420.868L34.7803 427.559L36.8677 420.532C39.7659 420.575 42.5754 421.537 44.8907 423.281C47.206 425.025 48.9071 427.46 49.7486 430.233C50.5901 433.007 50.5283 435.976 49.5723 438.712C48.6162 441.449 46.8152 443.81 44.4295 445.456C42.0437 447.102 39.1966 447.948 36.2991 447.87C33.4016 447.793 30.6038 446.796 28.3095 445.025C26.0152 443.254 24.3432 440.799 23.5349 438.016C22.7265 435.232 22.8237 432.264 23.8124 429.539V429.539Z"
+                fill="#57B894"
+              />
+              <path
+                d="M24.5632 427.123C25.3443 424.968 26.6552 423.044 28.3752 421.529C30.0952 420.013 32.169 418.955 34.4053 418.452L35.5312 425.143L37.6186 418.116C40.5168 418.158 43.3263 419.121 45.6415 420.865C47.9568 422.609 49.658 425.044 50.4995 427.817C51.3409 430.591 51.2792 433.56 50.3231 436.296C49.367 439.033 47.5661 441.394 45.1803 443.04C42.7946 444.686 39.9474 445.532 37.0499 445.454C34.1525 445.377 31.3546 444.38 29.0603 442.609C26.7661 440.838 25.094 438.383 24.2857 435.599C23.4774 432.816 23.5746 429.848 24.5632 427.123V427.123Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M34.0576 501.822C37.214 502.967 40.6813 502.91 43.7987 501.663C46.9162 500.416 49.4656 498.065 50.9612 495.059C52.4568 492.052 52.7939 488.601 51.9083 485.362C51.0227 482.124 48.9763 479.324 46.1592 477.497L44.5553 482.897L43.418 476.127C43.4067 476.122 43.3951 476.118 43.3832 476.114C41.6948 475.5 39.9021 475.225 38.1073 475.304C36.3126 475.384 34.551 475.816 32.9234 476.576C31.2957 477.336 29.8338 478.41 28.6213 479.736C27.4088 481.061 26.4693 482.613 25.8567 484.301C25.2441 485.99 24.9703 487.783 25.0509 489.578C25.1315 491.373 25.565 493.134 26.3266 494.761C27.0882 496.388 28.163 497.849 29.4895 499.061C30.816 500.272 32.3683 501.21 34.0576 501.822V501.822Z"
+                fill="#57B894"
+              />
+              <path
+                d="M34.7081 499.374C37.8645 500.519 41.3319 500.463 44.4493 499.216C47.5667 497.968 50.1162 495.618 51.6118 492.611C53.1074 489.605 53.4445 486.154 52.5589 482.915C51.6733 479.676 49.6269 476.877 46.8098 475.05L45.2059 480.45L44.0685 473.68C44.0573 473.675 44.0457 473.67 44.0337 473.667C42.3458 473.055 40.5537 472.781 38.7599 472.861C36.9661 472.941 35.2057 473.374 33.5791 474.135C30.2941 475.671 27.7539 478.449 26.5172 481.858C25.2806 485.267 25.4488 489.028 26.9849 492.313C28.521 495.597 31.2991 498.138 34.7081 499.374V499.374Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+              <path
+                d="M19.822 524.53C20.3647 527.618 21.9528 530.426 24.3198 532.482C26.6868 534.538 29.6889 535.718 32.8226 535.823C35.9562 535.929 39.0309 534.954 41.5309 533.061C44.0309 531.169 45.8042 528.475 46.5535 525.43C47.3027 522.385 46.9823 519.176 45.6459 516.339C44.3096 513.503 42.0386 511.212 39.2139 509.852C36.3891 508.491 33.1823 508.143 30.1314 508.866C27.0804 509.589 24.3708 511.339 22.457 513.822L28.5493 522.992L20.2249 518.125C19.5795 520.197 19.4413 522.393 19.822 524.53V524.53Z"
+                fill="#57B894"
+              />
+              <path
+                d="M20.3851 522.718C20.9278 525.806 22.5159 528.614 24.8829 530.67C27.25 532.726 30.252 533.906 33.3857 534.011C36.5194 534.117 39.594 533.142 42.094 531.249C44.594 529.357 46.3673 526.663 47.1166 523.618C47.8658 520.573 47.5454 517.364 46.2091 514.527C44.8727 511.691 42.6017 509.4 39.777 508.039C36.9522 506.679 33.7454 506.331 30.6945 507.054C27.6435 507.777 24.9339 509.527 23.0201 512.01L29.1124 521.18L20.788 516.313C20.1426 518.385 20.0044 520.581 20.3851 522.718V522.718Z"
+                className="text-gray-800 dark:text-gray-600"
+                stroke="currentColor"
+                strokeMiterlimit="10"
+              />
+            </g>
+            <defs>
+              <clipPath id="clip0">
+                <rect width="1074.39" height="584.231" fill="white" />
+              </clipPath>
+            </defs>
+          </Box>
+        </motion.div>
+
+        <motion.div
+          initial={{ opacity: 0, y: 40 }}
+          animate={{ opacity: 1, y: 0, transition: { delay: 0.2 } }}
+        >
+          <Typography
+            variant="h1"
+            className="mt-48 sm:mt-96 text-4xl md:text-7xl font-extrabold tracking-tight leading-tight md:leading-none text-center"
+          >
+            Ooops... 404!
           </Typography>
         </motion.div>
 
@@ -22,26 +235,17 @@ function Error404Page() {
           initial={{ opacity: 0, y: 40 }}
           animate={{ opacity: 1, y: 0, transition: { delay: 0.2 } }}
         >
-          <Typography variant="h5" color="textSecondary" className="mb-16 font-normal">
-            Sorry but we could not find the page you are looking for
+          <Typography
+            variant="h5"
+            color="text.secondary"
+            className="mt-8 text-lg md:text-xl font-medium tracking-tight text-center"
+          >
+            The page you requested could not be found.
           </Typography>
         </motion.div>
 
-        <Paper className="flex items-center w-full h-56 p-16 mt-48 mb-16 shadow">
-          <Icon color="action">search</Icon>
-          <Input
-            placeholder="Search for anything"
-            className="px-16"
-            disableUnderline
-            fullWidth
-            inputProps={{
-              'aria-label': 'Search',
-            }}
-          />
-        </Paper>
-
-        <Link className="font-normal" to="/">
-          Go back to dashboard
+        <Link className="block font-normal mt-48" to="/">
+          Back to Dashboard
         </Link>
       </div>
     </div>
