@@ -1,8 +1,9 @@
 import i18next from 'i18next';
-import Example from './Example';
+
 import en from './i18n/en';
 import tr from './i18n/tr';
 import ar from './i18n/ar';
+import Example from './Example';
 
 i18next.addResourceBundle('en', 'examplePage', en);
 i18next.addResourceBundle('tr', 'examplePage', tr);
@@ -14,7 +15,6 @@ const ExampleConfig = {
       config: {},
     },
   },
-  auth: ["admin"],
   routes: [
     {
       path: 'example',
@@ -30,22 +30,22 @@ export default ExampleConfig;
  */
 /*
 import React from 'react';
-const Example = React.lazy(() => import('./Example'));
+
+const Example = lazy(() => import('./Example'));
 
 const ExampleConfig = {
-    settings: {
-        layout: {
-            config: {}
-        }
+  settings: {
+    layout: {
+      config: {},
     },
-    routes  : [
-        {
-            path     : 'example',
-            element: <Example/>
-        }
-    ]
+  },
+  routes: [
+    {
+      path: 'example',
+      element: <Example />,
+    },
+  ],
 };
 
 export default ExampleConfig;
-
 */
