@@ -1,8 +1,11 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { styled } from '@mui/system';
 import FusePageSimple from '@fuse/core/FusePageSimple';
 import { useTranslation } from 'react-i18next';
 import { Button } from '@mui/material';
+import { useDispatch, useSelector } from 'react-redux';
+import { useParams } from 'react-router-dom';
+import { fetchHistoryTradingDetail } from './store/historyTradingSlice';
 import Edit from './Edit';
 import Detail from './Detail';
 
@@ -16,9 +19,17 @@ const Root = styled(FusePageSimple)({
   '& .FusePageSimple-sidebarContent': {},
 });
 
-export default function UnstyledTable() {
+export default function HistoryTradingDetail(props) {
   const [isEdit, setIsEdit] = useState(false);
   const { t } = useTranslation('historyTrading');
+  const params = useParams();
+  const goldLession = useSelector(({ historyTrading }) => historyTrading.goldLession);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchHistoryTradingDetail(params.id));
+  }, [dispatch, params.id]);
+
   return (
     <Root
       header={
@@ -26,12 +37,23 @@ export default function UnstyledTable() {
           <h4>{t('TITLE')}</h4>
         </div>
       }
-      contentToolbar={
-        <div className="px-24">
-          <Button onClick={() => setIsEdit(!isEdit)}>{isEdit ? 'Edit' : 'Detail'}</Button>
-        </div>
+      // contentToolbar={
+
+      // }
+      content={
+        <>
+          <div className="px-24">
+            <Button onClick={(e) => {
+              e.preventDefault();
+                setIsEdit(!isEdit);
+              }}
+            >
+              {isEdit ? 'Edit' : 'Detail'}
+            </Button>
+          </div>
+          {isEdit ? <Edit goldLession={goldLession} /> : <Detail goldLession={goldLession} />}
+        </>
       }
-      content={isEdit ? <Edit /> : <Detail />}
     />
   );
 }
