@@ -1,20 +1,20 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import auth from 'app/auth/store';
 import fuse from './fuse';
 import i18n from './i18nSlice';
+import user from './userSlice';
 
 const createReducer = (asyncReducers) => (state, action) => {
   const combinedReducer = combineReducers({
-    auth,
     fuse,
     i18n,
+    user,
     ...asyncReducers,
   });
 
   /*
 	Reset the redux store when user logged out
 	 */
-  if (action.type === 'auth/user/userLoggedOut') {
+  if (action.type === 'user/userLoggedOut') {
     // state = undefined;
   }
 
