@@ -4,12 +4,12 @@
 // import jssExtend from 'jss-plugin-extend';
 // import rtl from 'jss-rtl';
 import Provider from 'react-redux/es/components/Provider';
-import AdapterDateFns from '@mui/lab/AdapterDateFns';
-import LocalizationProvider from '@mui/lab/LocalizationProvider';
+import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { StyledEngineProvider } from '@mui/material/styles';
-import AppContext from './AppContext';
-import routes from './fuse-configs/routesConfig';
+import routes from 'app/configs/routesConfig';
 import store from './store';
+import AppContext from './AppContext';
 
 const withAppProviders = (Component) => (props) => {
   const WrapperComponent = () => (
