@@ -1,20 +1,180 @@
 /* eslint-disable import/no-extraneous-dependencies */
+const path = require('path');
+
 module.exports = {
   content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   safelist: ['pl-24', 'pl-40', 'pl-56', 'pl-72', 'pl-80'],
   presets: [],
-  darkMode: 'media', // or 'class'
+  darkMode: 'class', // or 'class'
   theme: {
-    screens: {
-      sm: '600px',
-      md: '960px',
-      lg: '1280px',
-      xl: '1920px',
-      print: { raw: 'print' },
+    accentColor: ({ theme }) => ({
+      ...theme('colors'),
+      auto: 'auto',
+    }),
+    animation: {
+      none: 'none',
+      spin: 'spin 1s linear infinite',
+      ping: 'ping 1s cubic-bezier(0, 0, 0.2, 1) infinite',
+      pulse: 'pulse 2s cubic-bezier(0.4, 0, 0.6, 1) infinite',
+      bounce: 'bounce 1s infinite',
     },
+    aria: {
+      checked: 'checked="true"',
+      disabled: 'disabled="true"',
+      expanded: 'expanded="true"',
+      hidden: 'hidden="true"',
+      pressed: 'pressed="true"',
+      readonly: 'readonly="true"',
+      required: 'required="true"',
+      selected: 'selected="true"',
+    },
+    aspectRatio: {
+      auto: 'auto',
+      square: '1 / 1',
+      video: '16 / 9',
+    },
+    backdropBlur: ({ theme }) => theme('blur'),
+    backdropBrightness: ({ theme }) => theme('brightness'),
+    backdropContrast: ({ theme }) => theme('contrast'),
+    backdropGrayscale: ({ theme }) => theme('grayscale'),
+    backdropHueRotate: ({ theme }) => theme('hueRotate'),
+    backdropInvert: ({ theme }) => theme('invert'),
+    backdropOpacity: ({ theme }) => theme('opacity'),
+    backdropSaturate: ({ theme }) => theme('saturate'),
+    backdropSepia: ({ theme }) => theme('sepia'),
+    backgroundColor: ({ theme }) => theme('colors'),
+    backgroundImage: {
+      none: 'none',
+      'gradient-to-t': 'linear-gradient(to top, var(--tw-gradient-stops))',
+      'gradient-to-tr': 'linear-gradient(to top right, var(--tw-gradient-stops))',
+      'gradient-to-r': 'linear-gradient(to right, var(--tw-gradient-stops))',
+      'gradient-to-br': 'linear-gradient(to bottom right, var(--tw-gradient-stops))',
+      'gradient-to-b': 'linear-gradient(to bottom, var(--tw-gradient-stops))',
+      'gradient-to-bl': 'linear-gradient(to bottom left, var(--tw-gradient-stops))',
+      'gradient-to-l': 'linear-gradient(to left, var(--tw-gradient-stops))',
+      'gradient-to-tl': 'linear-gradient(to top left, var(--tw-gradient-stops))',
+    },
+    backgroundOpacity: ({ theme }) => theme('opacity'),
+    backgroundPosition: {
+      bottom: 'bottom',
+      center: 'center',
+      left: 'left',
+      'left-bottom': 'left bottom',
+      'left-top': 'left top',
+      right: 'right',
+      'right-bottom': 'right bottom',
+      'right-top': 'right top',
+      top: 'top',
+    },
+    backgroundSize: {
+      auto: 'auto',
+      cover: 'cover',
+      contain: 'contain',
+    },
+    blur: {
+      0: '0',
+      none: '0',
+      sm: '4px',
+      DEFAULT: '8px',
+      md: '12px',
+      lg: '16px',
+      xl: '24px',
+      '2xl': '40px',
+      '3xl': '64px',
+    },
+    borderColor: ({ theme }) => ({
+      ...theme('colors'),
+      DEFAULT: theme('colors.gray.200', 'currentColor'),
+    }),
+    borderOpacity: ({ theme }) => theme('opacity'),
+    borderRadius: {
+      none: '0px',
+      sm: '.2rem',
+      DEFAULT: '.4rem',
+      md: '.6rem',
+      lg: '.8rem',
+      xl: '1.2rem',
+      '2xl': '1.6rem',
+      '3xl': '2.4rem',
+      full: '9999px',
+      0: '0px',
+      2: '.2rem',
+      4: '.4rem',
+      6: '.6rem',
+      8: '.8rem',
+      12: '1.2rem',
+      16: '1.6rem',
+      20: '2rem',
+      24: '2.4rem',
+      28: '2.8rem',
+      32: '3.2rem',
+    },
+    borderSpacing: ({ theme }) => ({
+      ...theme('spacing'),
+    }),
+    borderWidth: {
+      DEFAULT: '1px',
+      0: '0px',
+      1: '1px',
+      2: '2px',
+      3: '3px',
+      4: '4px',
+      8: '8px',
+    },
+    boxShadow: {
+      sm: '0 1px 2px 0 rgba(0, 0, 0, 0.05)',
+      DEFAULT: '0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06)',
+      md: '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
+      lg: '0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05)',
+      xl: '0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)',
+      '2xl': '0 25px 50px -12px rgba(0, 0, 0, 0.25)',
+      inner: 'inset 0 2px 4px 0 rgba(0,0,0,0.06)',
+      none: 'none',
+      0: '0px 0px 0px 0px rgba(0, 0, 0, 0.2), 0px 0px 0px 0px rgba(0, 0, 0, 0.14), 0px 0px 0px 0px rgba(0, 0, 0, 0.12)',
+      1: '0px 2px 1px -1px rgba(0, 0, 0, 0.2), 0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12)',
+      2: '0px 3px 1px -2px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 1px 5px 0px rgba(0, 0, 0, 0.12)',
+      3: '0px 3px 3px -2px rgba(0, 0, 0, 0.2), 0px 3px 4px 0px rgba(0, 0, 0, 0.14), 0px 1px 8px 0px rgba(0, 0, 0, 0.12)',
+      4: '0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12)',
+      5: '0px 3px 5px -1px rgba(0, 0, 0, 0.2), 0px 5px 8px 0px rgba(0, 0, 0, 0.14), 0px 1px 14px 0px rgba(0, 0, 0, 0.12)',
+      6: '0px 3px 5px -1px rgba(0, 0, 0, 0.2), 0px 6px 10px 0px rgba(0, 0, 0, 0.14), 0px 1px 18px 0px rgba(0, 0, 0, 0.12)',
+      7: '0px 4px 5px -2px rgba(0, 0, 0, 0.2), 0px 7px 10px 1px rgba(0, 0, 0, 0.14), 0px 2px 16px 1px rgba(0, 0, 0, 0.12)',
+      8: '0px 5px 5px -3px rgba(0, 0, 0, 0.2), 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12)',
+      9: '0px 5px 6px -3px rgba(0, 0, 0, 0.2), 0px 9px 12px 1px rgba(0, 0, 0, 0.14), 0px 3px 16px 2px rgba(0, 0, 0, 0.12)',
+      10: '0px 6px 6px -3px rgba(0, 0, 0, 0.2), 0px 10px 14px 1px rgba(0, 0, 0, 0.14), 0px 4px 18px 3px rgba(0, 0, 0, 0.12)',
+      11: '0px 6px 7px -4px rgba(0, 0, 0, 0.2), 0px 11px 15px 1px rgba(0, 0, 0, 0.14), 0px 4px 20px 3px rgba(0, 0, 0, 0.12)',
+      12: '0px 7px 8px -4px rgba(0, 0, 0, 0.2), 0px 12px 17px 2px rgba(0, 0, 0, 0.14), 0px 5px 22px 4px rgba(0, 0, 0, 0.12)',
+      13: '0px 7px 8px -4px rgba(0, 0, 0, 0.2), 0px 13px 19px 2px rgba(0, 0, 0, 0.14), 0px 5px 24px 4px rgba(0, 0, 0, 0.12)',
+      14: '0px 7px 9px -4px rgba(0, 0, 0, 0.2), 0px 14px 21px 2px rgba(0, 0, 0, 0.14), 0px 5px 26px 4px rgba(0, 0, 0, 0.12)',
+      15: '0px 8px 9px -5px rgba(0, 0, 0, 0.2), 0px 15px 22px 2px rgba(0, 0, 0, 0.14), 0px 6px 28px 5px rgba(0, 0, 0, 0.12)',
+      16: '0px 8px 10px -5px rgba(0, 0, 0, 0.2), 0px 16px 24px 2px rgba(0, 0, 0, 0.14), 0px 6px 30px 5px rgba(0, 0, 0, 0.12)',
+      17: '0px 8px 11px -5px rgba(0, 0, 0, 0.2), 0px 17px 26px 2px rgba(0, 0, 0, 0.14), 0px 6px 32px 5px rgba(0, 0, 0, 0.12)',
+      18: '0px 9px 11px -5px rgba(0, 0, 0, 0.2), 0px 18px 28px 2px rgba(0, 0, 0, 0.14), 0px 7px 34px 6px rgba(0, 0, 0, 0.12)',
+      19: '0px 9px 12px -6px rgba(0, 0, 0, 0.2), 0px 19px 29px 2px rgba(0, 0, 0, 0.14), 0px 7px 36px 6px rgba(0, 0, 0, 0.12)',
+      20: '0px 10px 13px -6px rgba(0, 0, 0, 0.2), 0px 20px 31px 3px rgba(0, 0, 0, 0.14), 0px 8px 38px 7px rgba(0, 0, 0, 0.12)',
+      21: '0px 10px 13px -6px rgba(0, 0, 0, 0.2), 0px 21px 33px 3px rgba(0, 0, 0, 0.14), 0px 8px 40px 7px rgba(0, 0, 0, 0.12)',
+      22: '0px 10px 14px -6px rgba(0, 0, 0, 0.2), 0px 22px 35px 3px rgba(0, 0, 0, 0.14), 0px 8px 42px 7px rgba(0, 0, 0, 0.12)',
+      23: '0px 11px 14px -7px rgba(0, 0, 0, 0.2), 0px 23px 36px 3px rgba(0, 0, 0, 0.14), 0px 9px 44px 8px rgba(0, 0, 0, 0.12)',
+      24: '0px 11px 15px -7px rgba(0, 0, 0, 0.2), 0px 24px 38px 3px rgba(0, 0, 0, 0.14), 0px 9px 46px 8px rgba(0, 0, 0, 0.12)',
+    },
+    boxShadowColor: ({ theme }) => theme('colors'),
+    brightness: {
+      0: '0',
+      50: '.5',
+      75: '.75',
+      90: '.9',
+      95: '.95',
+      100: '1',
+      105: '1.05',
+      110: '1.1',
+      125: '1.25',
+      150: '1.5',
+      200: '2',
+    },
+    caretColor: ({ theme }) => theme('colors'),
     colors: ({ colors }) => ({
+      inherit: colors.inherit,
+      current: colors.current,
       transparent: 'transparent',
-      current: 'currentColor',
       black: '#22292F',
       white: '#fff',
       grey: {
@@ -374,235 +534,22 @@ module.exports = {
       12: '12',
       '3xs': '16rem',
       '2xs': '18rem',
-      xs: '20rem',
-      sm: '24rem',
-      md: '28rem',
-      lg: '32rem',
-      xl: '36rem',
-      '2xl': '42rem',
-      '3xl': '48rem',
-      '4xl': '56rem',
-      '5xl': '64rem',
-      '6xl': '72rem',
-      '7xl': '80rem',
-    },
-    spacing: {
       xs: '32rem',
       sm: '48rem',
       md: '64rem',
       lg: '80rem',
-      xl: '96rem',
-      '2xl': '112rem',
-      '3xl': '128rem',
-      '4xl': '144rem',
-      '5xl': '160rem',
-      px: '1px',
-      0: '0',
-      0.5: '0.05rem',
-      1: '0.1rem',
-      1.5: '0.15rem',
-      2: '0.2rem',
-      2.5: '0.25rem',
-      3: '0.3rem',
-      3.5: '0.35rem',
-      4: '0.4rem',
-      5: '0.5rem',
-      6: '0.6rem',
-      7: '0.7rem',
-      8: '0.8rem',
-      9: '0.9rem',
-      10: '1.0rem',
-      11: '1.1rem',
-      12: '1.2rem',
-      14: '1.4rem',
-      16: '1.6rem',
-      20: '2rem',
-      24: '2.4rem',
-      28: '2.8rem',
-      32: '3.2rem',
-      36: '3.6rem',
-      40: '4rem',
-      44: '4.4rem',
-      48: '4.8rem',
-      52: '5.2rem',
-      56: '5.6rem',
-      60: '6rem',
-      64: '6.4rem',
-      68: '6.8rem',
-      72: '7.2rem',
-      76: '7.6rem',
-      80: '8rem',
-      84: '8.4rem',
-      88: '8.8rem',
-      92: '9.2rem',
-      96: '9.6rem',
-      128: '12.8rem',
-      136: '13.6rem',
-      160: '16rem',
-      192: '19.2rem',
-      200: '20rem',
-      208: '20.8rem',
-      216: '21.6rem',
-      224: '22.4rem',
-      256: '25.6rem',
-      288: '28.8rem',
-      320: '32rem',
-      360: '36rem',
-      384: '38.4rem',
-      400: '40rem',
-      512: '51.2rem',
-      640: '64rem',
+      xl: '57.6rem',
+      '2xl': '65.6rem',
+      '3xl': '76.8rem',
+      '4xl': '89.6rem',
+      '5xl': '102.4rem',
+      '6xl': '115.2rem',
+      '7xl': '128rem',
     },
-    animation: {
+    container: {},
+    content: {
       none: 'none',
-      spin: 'spin 1s linear infinite',
-      ping: 'ping 1s cubic-bezier(0, 0, 0.2, 1) infinite',
-      pulse: 'pulse 2s cubic-bezier(0.4, 0, 0.6, 1) infinite',
-      bounce: 'bounce 1s infinite',
     },
-    aspectRatio: {
-      auto: 'auto',
-      square: '1 / 1',
-      video: '16 / 9',
-    },
-    backdropBlur: ({ theme }) => theme('blur'),
-    backdropBrightness: ({ theme }) => theme('brightness'),
-    backdropContrast: ({ theme }) => theme('contrast'),
-    backdropGrayscale: ({ theme }) => theme('grayscale'),
-    backdropHueRotate: ({ theme }) => theme('hueRotate'),
-    backdropInvert: ({ theme }) => theme('invert'),
-    backdropOpacity: ({ theme }) => theme('opacity'),
-    backdropSaturate: ({ theme }) => theme('saturate'),
-    backdropSepia: ({ theme }) => theme('sepia'),
-    backgroundColor: ({ theme }) => theme('colors'),
-    backgroundImage: {
-      none: 'none',
-      'gradient-to-t': 'linear-gradient(to top, var(--tw-gradient-stops))',
-      'gradient-to-tr': 'linear-gradient(to top right, var(--tw-gradient-stops))',
-      'gradient-to-r': 'linear-gradient(to right, var(--tw-gradient-stops))',
-      'gradient-to-br': 'linear-gradient(to bottom right, var(--tw-gradient-stops))',
-      'gradient-to-b': 'linear-gradient(to bottom, var(--tw-gradient-stops))',
-      'gradient-to-bl': 'linear-gradient(to bottom left, var(--tw-gradient-stops))',
-      'gradient-to-l': 'linear-gradient(to left, var(--tw-gradient-stops))',
-      'gradient-to-tl': 'linear-gradient(to top left, var(--tw-gradient-stops))',
-    },
-    backgroundOpacity: ({ theme }) => theme('opacity'),
-    backgroundPosition: {
-      bottom: 'bottom',
-      center: 'center',
-      left: 'left',
-      'left-bottom': 'left bottom',
-      'left-top': 'left top',
-      right: 'right',
-      'right-bottom': 'right bottom',
-      'right-top': 'right top',
-      top: 'top',
-    },
-    backgroundSize: {
-      auto: 'auto',
-      cover: 'cover',
-      contain: 'contain',
-    },
-    blur: {
-      0: '0',
-      none: '0',
-      sm: '4px',
-      DEFAULT: '8px',
-      md: '12px',
-      lg: '16px',
-      xl: '24px',
-      '2xl': '40px',
-      '3xl': '64px',
-    },
-    brightness: {
-      0: '0',
-      50: '.5',
-      75: '.75',
-      90: '.9',
-      95: '.95',
-      100: '1',
-      105: '1.05',
-      110: '1.1',
-      125: '1.25',
-      150: '1.5',
-      200: '2',
-    },
-    borderColor: ({ theme }) => ({
-      ...theme('colors'),
-      DEFAULT: theme('colors.gray.200', 'currentColor'),
-    }),
-    borderOpacity: ({ theme }) => theme('opacity'),
-    borderRadius: {
-      none: '0px',
-      sm: '.2rem',
-      DEFAULT: '.4rem',
-      md: '.6rem',
-      lg: '.8rem',
-      xl: '1.2rem',
-      '2xl': '1.6rem',
-      '3xl': '2.4rem',
-      full: '9999px',
-      2: '.2rem',
-      4: '.4rem',
-      6: '.6rem',
-      8: '.8rem',
-      12: '1.2rem',
-      16: '1.6rem',
-      20: '2rem',
-      24: '2.4rem',
-      28: '2.8rem',
-      32: '3.2rem',
-    },
-    borderWidth: {
-      DEFAULT: '1px',
-      0: '0px',
-      1: '1px',
-      2: '2px',
-      3: '3px',
-      4: '4px',
-      8: '8px',
-    },
-    boxShadow: {
-      sm: '0 1px 2px 0 rgba(0, 0, 0, 0.05)',
-      DEFAULT: '0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06)',
-      md: '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
-      lg: '0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05)',
-      xl: '0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)',
-      '2xl': '0 25px 50px -12px rgba(0, 0, 0, 0.25)',
-      inner: 'inset 0 2px 4px 0 rgba(0,0,0,0.06)',
-      none: 'none',
-      0: '0px 0px 0px 0px rgba(0, 0, 0, 0.2), 0px 0px 0px 0px rgba(0, 0, 0, 0.14), 0px 0px 0px 0px rgba(0, 0, 0, 0.12)',
-      1: '0px 2px 1px -1px rgba(0, 0, 0, 0.2), 0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12)',
-      2: '0px 3px 1px -2px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 1px 5px 0px rgba(0, 0, 0, 0.12)',
-      3: '0px 3px 3px -2px rgba(0, 0, 0, 0.2), 0px 3px 4px 0px rgba(0, 0, 0, 0.14), 0px 1px 8px 0px rgba(0, 0, 0, 0.12)',
-      4: '0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12)',
-      5: '0px 3px 5px -1px rgba(0, 0, 0, 0.2), 0px 5px 8px 0px rgba(0, 0, 0, 0.14), 0px 1px 14px 0px rgba(0, 0, 0, 0.12)',
-      6: '0px 3px 5px -1px rgba(0, 0, 0, 0.2), 0px 6px 10px 0px rgba(0, 0, 0, 0.14), 0px 1px 18px 0px rgba(0, 0, 0, 0.12)',
-      7: '0px 4px 5px -2px rgba(0, 0, 0, 0.2), 0px 7px 10px 1px rgba(0, 0, 0, 0.14), 0px 2px 16px 1px rgba(0, 0, 0, 0.12)',
-      8: '0px 5px 5px -3px rgba(0, 0, 0, 0.2), 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12)',
-      9: '0px 5px 6px -3px rgba(0, 0, 0, 0.2), 0px 9px 12px 1px rgba(0, 0, 0, 0.14), 0px 3px 16px 2px rgba(0, 0, 0, 0.12)',
-      10: '0px 6px 6px -3px rgba(0, 0, 0, 0.2), 0px 10px 14px 1px rgba(0, 0, 0, 0.14), 0px 4px 18px 3px rgba(0, 0, 0, 0.12)',
-      11: '0px 6px 7px -4px rgba(0, 0, 0, 0.2), 0px 11px 15px 1px rgba(0, 0, 0, 0.14), 0px 4px 20px 3px rgba(0, 0, 0, 0.12)',
-      12: '0px 7px 8px -4px rgba(0, 0, 0, 0.2), 0px 12px 17px 2px rgba(0, 0, 0, 0.14), 0px 5px 22px 4px rgba(0, 0, 0, 0.12)',
-      13: '0px 7px 8px -4px rgba(0, 0, 0, 0.2), 0px 13px 19px 2px rgba(0, 0, 0, 0.14), 0px 5px 24px 4px rgba(0, 0, 0, 0.12)',
-      14: '0px 7px 9px -4px rgba(0, 0, 0, 0.2), 0px 14px 21px 2px rgba(0, 0, 0, 0.14), 0px 5px 26px 4px rgba(0, 0, 0, 0.12)',
-      15: '0px 8px 9px -5px rgba(0, 0, 0, 0.2), 0px 15px 22px 2px rgba(0, 0, 0, 0.14), 0px 6px 28px 5px rgba(0, 0, 0, 0.12)',
-      16: '0px 8px 10px -5px rgba(0, 0, 0, 0.2), 0px 16px 24px 2px rgba(0, 0, 0, 0.14), 0px 6px 30px 5px rgba(0, 0, 0, 0.12)',
-      17: '0px 8px 11px -5px rgba(0, 0, 0, 0.2), 0px 17px 26px 2px rgba(0, 0, 0, 0.14), 0px 6px 32px 5px rgba(0, 0, 0, 0.12)',
-      18: '0px 9px 11px -5px rgba(0, 0, 0, 0.2), 0px 18px 28px 2px rgba(0, 0, 0, 0.14), 0px 7px 34px 6px rgba(0, 0, 0, 0.12)',
-      19: '0px 9px 12px -6px rgba(0, 0, 0, 0.2), 0px 19px 29px 2px rgba(0, 0, 0, 0.14), 0px 7px 36px 6px rgba(0, 0, 0, 0.12)',
-      20: '0px 10px 13px -6px rgba(0, 0, 0, 0.2), 0px 20px 31px 3px rgba(0, 0, 0, 0.14), 0px 8px 38px 7px rgba(0, 0, 0, 0.12)',
-      21: '0px 10px 13px -6px rgba(0, 0, 0, 0.2), 0px 21px 33px 3px rgba(0, 0, 0, 0.14), 0px 8px 40px 7px rgba(0, 0, 0, 0.12)',
-      22: '0px 10px 14px -6px rgba(0, 0, 0, 0.2), 0px 22px 35px 3px rgba(0, 0, 0, 0.14), 0px 8px 42px 7px rgba(0, 0, 0, 0.12)',
-      23: '0px 11px 14px -7px rgba(0, 0, 0, 0.2), 0px 23px 36px 3px rgba(0, 0, 0, 0.14), 0px 9px 44px 8px rgba(0, 0, 0, 0.12)',
-      24: '0px 11px 15px -7px rgba(0, 0, 0, 0.2), 0px 24px 38px 3px rgba(0, 0, 0, 0.14), 0px 9px 46px 8px rgba(0, 0, 0, 0.12)',
-    },
-    boxShadowColor: ({ theme }) => theme('colors'),
-    caretColor: ({ theme }) => theme('colors'),
-    accentColor: ({ theme }) => ({
-      ...theme('colors'),
-      auto: 'auto',
-    }),
     contrast: {
       0: '0',
       50: '.5',
@@ -611,10 +558,6 @@ module.exports = {
       125: '1.25',
       150: '1.5',
       200: '2',
-    },
-    container: {},
-    content: {
-      none: 'none',
     },
     cursor: {
       auto: 'auto',
@@ -666,28 +609,10 @@ module.exports = {
       '2xl': '0 25px 25px rgba(0, 0, 0, 0.15)',
       none: '0 0 #0000',
     },
-    fill: ({ theme }) => theme('colors'),
-    grayscale: {
-      0: '0',
-      DEFAULT: '100%',
-    },
-    hueRotate: {
-      '-180': '-180deg',
-      '-90': '-90deg',
-      '-60': '-60deg',
-      '-30': '-30deg',
-      '-15': '-15deg',
-      0: '0deg',
-      15: '15deg',
-      30: '30deg',
-      60: '60deg',
-      90: '90deg',
-      180: '180deg',
-    },
-    invert: {
-      0: '0',
-      DEFAULT: '100%',
-    },
+    fill: ({ theme }) => ({
+      none: 'none',
+      ...theme('colors'),
+    }),
     flex: {
       1: '1 1 0%',
       auto: '1 1 auto',
@@ -735,7 +660,7 @@ module.exports = {
     },
     fontFamily: {
       sans: [
-        'Muli',
+        'Inter var',
         'Roboto',
         '-apple-system',
         'BlinkMacSystemFont',
@@ -750,23 +675,34 @@ module.exports = {
         '"Segoe UI Symbol"',
         '"Noto Color Emoji"',
       ],
-      serif: ['Georgia', 'Cambria', '"Times New Roman"', 'Times', 'serif'],
-      mono: ['Consolas', '"Liberation Mono"', 'Menlo', 'Courier', 'monospace'],
+      serif: ['ui-serif', 'Georgia', 'Cambria', '"Times New Roman"', 'Times', 'serif'],
+      mono: [
+        'ui-monospace',
+        'SFMono-Regular',
+        'Menlo',
+        'Monaco',
+        'Consolas',
+        '"Liberation Mono"',
+        '"Courier New"',
+        'monospace',
+      ],
     },
     fontSize: {
-      xs: ['1.2rem', { lineHeight: '1.6rem' }],
-      sm: ['1.4rem', { lineHeight: '2rem' }],
-      base: ['1.6rem', { lineHeight: '2.4rem' }],
-      lg: ['1.8rem', { lineHeight: '2.8em' }],
-      xl: ['2rem', { lineHeight: '2.8rem' }],
-      '2xl': ['2.4rem', { lineHeight: '3.2rem' }],
-      '3xl': ['3rem', { lineHeight: '3.6rem' }],
-      '4xl': ['3.6rem', { lineHeight: '4rem' }],
-      '5xl': ['4.8rem', { lineHeight: '1' }],
-      '6xl': ['6rem', { lineHeight: '1' }],
-      '7xl': ['7.2rem', { lineHeight: '1' }],
-      '8xl': ['9.6rem', { lineHeight: '1' }],
-      '9xl': ['12.8rem', { lineHeight: '1' }],
+      xs: '1rem',
+      sm: '1.2rem',
+      md: '1.3rem',
+      base: '1.4rem',
+      lg: '1.6rem',
+      xl: '1.8rem',
+      '2xl': '2rem',
+      '3xl': '2.4rem',
+      '4xl': '3.2rem',
+      '5xl': '3.6rem',
+      '6xl': '4rem',
+      '7xl': '4.8rem',
+      '8xl': '6.4rem',
+      '9xl': '9.6rem',
+      '10xl': '12.8rem',
       10: ['1rem'],
       11: ['1.1rem'],
       12: ['1.2rem'],
@@ -816,6 +752,33 @@ module.exports = {
     },
     gap: ({ theme }) => theme('spacing'),
     gradientColorStops: ({ theme }) => theme('colors'),
+    gradientColorStopPositions: {
+      '0%': '0%',
+      '5%': '5%',
+      '10%': '10%',
+      '15%': '15%',
+      '20%': '20%',
+      '25%': '25%',
+      '30%': '30%',
+      '35%': '35%',
+      '40%': '40%',
+      '45%': '45%',
+      '50%': '50%',
+      '55%': '55%',
+      '60%': '60%',
+      '65%': '65%',
+      '70%': '70%',
+      '75%': '75%',
+      '80%': '80%',
+      '85%': '85%',
+      '90%': '90%',
+      '95%': '95%',
+      '100%': '100%',
+    },
+    grayscale: {
+      0: '0',
+      DEFAULT: '100%',
+    },
     gridAutoColumns: {
       auto: 'auto',
       min: 'min-content',
@@ -886,7 +849,7 @@ module.exports = {
       'span-6': 'span 6 / span 6',
       'span-full': '1 / -1',
     },
-    gridRowStart: {
+    gridRowEnd: {
       auto: 'auto',
       1: '1',
       2: '2',
@@ -896,7 +859,7 @@ module.exports = {
       6: '6',
       7: '7',
     },
-    gridRowEnd: {
+    gridRowStart: {
       auto: 'auto',
       1: '1',
       2: '2',
@@ -954,6 +917,19 @@ module.exports = {
       max: 'max-content',
       fit: 'fit-content',
     }),
+    hueRotate: {
+      '-180': '-180deg',
+      '-90': '-90deg',
+      '-60': '-60deg',
+      '-30': '-30deg',
+      '-15': '-15deg',
+      0: '0deg',
+      15: '15deg',
+      30: '30deg',
+      60: '60deg',
+      90: '90deg',
+      180: '180deg',
+    },
     inset: (theme, { negative }) => ({
       auto: 'auto',
       ...theme('spacing'),
@@ -973,6 +949,10 @@ module.exports = {
       '-3/4': '-75%',
       '-full': '-100%',
     }),
+    invert: {
+      0: '0',
+      DEFAULT: '100%',
+    },
     keyframes: {
       spin: {
         to: {
@@ -1030,12 +1010,24 @@ module.exports = {
       disc: 'disc',
       decimal: 'decimal',
     },
+    listStyleImage: {
+      none: 'none',
+    },
     margin: (theme, { negative }) => ({
       auto: 'auto',
       ...theme('spacing'),
       ...negative(theme('spacing')),
     }),
+    lineClamp: {
+      1: '1',
+      2: '2',
+      3: '3',
+      4: '4',
+      5: '5',
+      6: '6',
+    },
     maxHeight: ({ theme }) => ({
+      none: 'none',
       ...theme('spacing'),
       full: '100%',
       screen: '100vh',
@@ -1059,6 +1051,9 @@ module.exports = {
       ...theme('spacing'),
       full: '100%',
       screen: '100vh',
+      min: 'min-content',
+      max: 'max-content',
+      fit: 'fit-content',
     }),
     minWidth: ({ theme }) => ({
       ...theme('spacing'),
@@ -1113,9 +1108,6 @@ module.exports = {
       11: '11',
       12: '12',
     },
-    padding: ({ theme }) => theme('spacing'),
-    placeholderColor: ({ theme }) => theme('colors'),
-    placeholderOpacity: ({ theme }) => theme('opacity'),
     outlineColor: ({ theme }) => theme('colors'),
     outlineOffset: {
       0: '0px',
@@ -1131,6 +1123,9 @@ module.exports = {
       4: '4px',
       8: '8px',
     },
+    padding: ({ theme }) => theme('spacing'),
+    placeholderColor: ({ theme }) => theme('colors'),
+    placeholderOpacity: ({ theme }) => theme('opacity'),
     ringColor: ({ theme }) => ({
       DEFAULT: theme('colors.blue.500', '#3b82f6'),
       ...theme('colors'),
@@ -1193,6 +1188,13 @@ module.exports = {
       125: '1.25',
       150: '1.5',
     },
+    screens: {
+      sm: '600px',
+      md: '960px',
+      lg: '1280px',
+      xl: '1920px',
+      print: { raw: 'print' },
+    },
     scrollMargin: ({ theme }) => ({
       ...theme('spacing'),
     }),
@@ -1218,12 +1220,90 @@ module.exports = {
       ...theme('spacing'),
       ...negative(theme('spacing')),
     }),
-    stroke: ({ theme }) => theme('colors'),
+    spacing: {
+      xs: '32rem',
+      sm: '48rem',
+      md: '64rem',
+      lg: '80rem',
+      xl: '96rem',
+      '2xl': '65.6rem',
+      '3xl': '76.8rem',
+      '4xl': '89.6rem',
+      '5xl': '102.4rem',
+      '6xl': '115.2rem',
+      '7xl': '128rem',
+      px: '1px',
+      0: '0px',
+      0.5: '0.05rem',
+      1: '0.1rem',
+      1.5: '0.15rem',
+      2: '0.2rem',
+      2.5: '0.25rem',
+      3: '0.3rem',
+      3.5: '0.35rem',
+      4: '0.4rem',
+      5: '0.5rem',
+      6: '0.6rem',
+      7: '0.7rem',
+      8: '0.8rem',
+      9: '0.9rem',
+      10: '1.0rem',
+      11: '1.1rem',
+      12: '1.2rem',
+      14: '1.4rem',
+      16: '1.6rem',
+      20: '2rem',
+      24: '2.4rem',
+      28: '2.8rem',
+      32: '3.2rem',
+      36: '3.6rem',
+      40: '4rem',
+      44: '4.4rem',
+      48: '4.8rem',
+      52: '5.2rem',
+      56: '5.6rem',
+      60: '6rem',
+      64: '6.4rem',
+      68: '6.8rem',
+      72: '7.2rem',
+      76: '7.6rem',
+      80: '8rem',
+      84: '8.4rem',
+      88: '8.8rem',
+      92: '9.2rem',
+      96: '9.6rem',
+      112: '11.2rem',
+      120: '12rem',
+      128: '12.8rem',
+      136: '13.6rem',
+      144: '14.4rem',
+      160: '16rem',
+      192: '19.2rem',
+      200: '20rem',
+      208: '20.8rem',
+      216: '21.6rem',
+      224: '22.4rem',
+      256: '25.6rem',
+      288: '28.8rem',
+      320: '32rem',
+      360: '36rem',
+      384: '38.4rem',
+      400: '40rem',
+      480: '48rem',
+      512: '51.2rem',
+      640: '64rem',
+    },
+    stroke: ({ theme }) => ({
+      none: 'none',
+      ...theme('colors'),
+    }),
     strokeWidth: {
       0: '0',
       1: '1',
       2: '2',
     },
+    supports: {},
+    data: {},
     textColor: ({ theme }) => theme('colors'),
     textDecorationColor: ({ theme }) => theme('colors'),
     textDecorationThickness: {
@@ -1235,6 +1315,10 @@ module.exports = {
       4: '4px',
       8: '8px',
     },
+    textIndent: ({ theme }) => ({
+      ...theme('spacing'),
+    }),
+    textOpacity: ({ theme }) => theme('opacity'),
     textUnderlineOffset: {
       auto: 'auto',
       0: '0px',
@@ -1243,10 +1327,6 @@ module.exports = {
       4: '4px',
       8: '8px',
     },
-    textIndent: ({ theme }) => ({
-      ...theme('spacing'),
-    }),
-    textOpacity: ({ theme }) => theme('opacity'),
     transformOrigin: {
       center: 'center',
       top: 'top',
@@ -1259,6 +1339,7 @@ module.exports = {
       'top-left': 'top left',
     },
     transitionDelay: {
+      0: '0s',
       75: '75ms',
       100: '100ms',
       150: '150ms',
@@ -1270,6 +1351,7 @@ module.exports = {
     },
     transitionDuration: {
       DEFAULT: '150ms',
+      0: '0s',
       75: '75ms',
       100: '100ms',
       150: '150ms',
@@ -1283,8 +1365,8 @@ module.exports = {
       none: 'none',
       all: 'all',
       DEFAULT:
-        'background-color, border-color, color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter',
-      colors: 'background-color, border-color, color, fill, stroke',
+        'color, background-color, border-color, text-decoration-color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter',
+      colors: 'color, background-color, border-color, text-decoration-color, fill, stroke',
       opacity: 'opacity',
       shadow: 'box-shadow',
       transform: 'transform',
@@ -1357,6 +1439,7 @@ module.exports = {
     },
     zIndex: {
       auto: 'auto',
+      '-1': -1,
       0: '0',
       10: '10',
       20: '20',
@@ -1366,6 +1449,26 @@ module.exports = {
       99: '99',
       999: '999',
       9999: '9999',
+    },
+    extend: {
+      // @tailwindcss/typography
+      typography: ({ theme }) => ({
+        DEFAULT: {
+          css: {
+            fontSize: '1.4rem',
+          },
+        },
+        sm: {
+          css: {
+            fontSize: '1.2rem',
+          },
+        },
+        lg: {
+          css: {
+            fontSize: '1.6rem',
+          },
+        },
+      }),
     },
   },
   variantOrder: [
@@ -1386,5 +1489,11 @@ module.exports = {
     'active',
     'disabled',
   ],
-  plugins: [require('tailwindcss-dir')()],
+  plugins: [
+    // eslint-disable-next-line import/no-dynamic-require
+    require(path.resolve(__dirname, 'src/@fuse/tailwind/plugins/icon-size')),
+    // Other third party and/or custom plugins
+    require('@tailwindcss/typography')({ modifiers: ['sm', 'lg'] }),
+    require('@tailwindcss/aspect-ratio'),
+  ],
 };
